% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.database.entries.R
\name{add.database.entries}
\alias{add.database.entries}
\title{Return new arrangement of database while adding code to deal with ensembles}
\usage{
add.database.entries(
  result,
  con,
  start_date,
  end_date,
  write,
  overwrite,
  insert.new.file,
  input.args,
  machine,
  mimetype,
  formatname,
  allow.conflicting.dates,
  ensemble,
  ensemble_name,
  existing.input,
  existing.dbfile,
  input
)
}
\arguments{
\item{result}{list of results from the download function}

\item{con}{database connection}

\item{start_date}{start date of the data}

\item{end_date}{end date of the data}

\item{write}{whether to write to the database}

\item{overwrite}{Logical: If a file already exists, create a fresh copy?}

\item{insert.new.file}{whether to insert a new file}

\item{input.args}{input arguments obtained from the convert_input function}

\item{machine}{machine information}

\item{mimetype}{data product specific file format}

\item{formatname}{format name of the data}

\item{allow.conflicting.dates}{whether to allow conflicting dates}

\item{ensemble}{ensemble id}

\item{ensemble_name}{ensemble name}

\item{existing.input}{existing input records}

\item{existing.dbfile}{existing dbfile records}

\item{input}{input records}
}
\value{
list of input and dbfile ids
}
\description{
Return new arrangement of database while adding code to deal with ensembles
}
\author{
Betsy Cowdery, Michael Dietze, Ankur Desai, Tony Gardella, Luke Dramko
}
