% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.treatments.R
\name{assign.treatments}
\alias{assign.treatments}
\title{assign.treatments}
\usage{
assign.treatments(data)
}
\arguments{
\item{data}{input data}
}
\value{
dataframe with sequential treatments
}
\description{
Change treatments to sequential integers
}
\details{
Assigns all control treatments the same value, then assigns unique treatments
within each site. Each site is required to have a control treatment.
The algorithm (incorrectly) assumes that each site has a unique set of experimental
treatments. This assumption is required by the data in BETYdb that does not always consistently name treatments or quantity them in the managements table. Also it avoids having the need to estimate treatment by site interactions in the meta analysis model. This model uses data in the control treatment to estimate model parameters so the impact of the assumption is minimal.
}
\author{
David LeBauer, Carl Davidson, Alexey Shiklomanov
}
