% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.format.vars.R
\name{bety2pecan}
\alias{bety2pecan}
\title{Convert BETY variable names to MsTMIP and subsequently PEcAn standard names}
\usage{
bety2pecan(vars_bety)
}
\arguments{
\item{vars_bety}{data frame with variable names and units}
}
\description{
Convert BETY variable names to MsTMIP and subsequently PEcAn standard names
}
\author{
Betsy Cowdery
}
