% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.dplyr.R
\name{betyConnect}
\alias{betyConnect}
\title{Connect to bety using current PEcAn configuration}
\usage{
betyConnect(php.config = "../../web/config.php")
}
\arguments{
\item{php.config}{Path to `config.php`}
}
\description{
Connect to bety using current PEcAn configuration
}
