% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.missing.files.R
\name{check_missing_files}
\alias{check_missing_files}
\title{Function to check if result has empty or missing files}
\usage{
check_missing_files(
  result,
  outname,
  existing.input = NULL,
  existing.dbfile = NULL
)
}
\arguments{
\item{result}{A list of dataframes with file paths}

\item{outname}{Name of the output file}

\item{existing.input}{Existing input records}

\item{existing.dbfile}{Existing dbfile records}
}
\value{
A list of dataframes with file paths, a list of strings with the output file name, a list of existing input records, and a list of existing dbfile records
}
\description{
Function to check if result has empty or missing files
}
\author{
Betsy Cowdery, Michael Dietze, Ankur Desai, Tony Gardella, Luke Dramko
}
