% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone_pft.R
\name{clone_pft}
\alias{clone_pft}
\title{Duplicate existing pft with associated priors, species, and cultivars}
\usage{
clone_pft(parent.pft.name, new.pft.name, new.pft.definition, settings)
}
\arguments{
\item{parent.pft.name}{name of PFT to duplicate}

\item{new.pft.name}{name for new PFT. Must not be the same as parent.pft.name}

\item{new.pft.definition}{text for the new PFT's definition field.}

\item{settings}{PEcAn settings list, used only for BETYdb connection parameters}
}
\value{
ID of the newly created pft in database, creates new PFT as a side effect
}
\description{
Creates a new pft that is a duplicate of an existing pft,
including relationships with priors, species, and cultivars (if any) of the existing pft.
This function mimics the 'clone pft' button in the PFTs record view page in the 
BETYdb web interface for PFTs that aggregate >=1 species, but adds the ability to 
clone the cultivar associations.
}
\examples{
\dontrun{
clone_pft(parent.pft.name    = "tempdecid",
          new.pft.name       = "mytempdecid",
          new.pft.definition = "mytempdecid is a new pft",
          settings = pecan_settings_list)
}
}
\author{
David LeBauer, Chris Black
}
