% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\docType{class}
\name{database-class}
\alias{database-class}
\alias{Db}
\title{An S4 class to wrap a SQLite database connection}
\description{
An S4 class to wrap a SQLite database connection
}
\details{
If the database file exists it is opened; otherwise, a new
database file is created. An exception is when the database file is
either ":memory:" or "". In the former case, a new database is created
in computer memory; in the latter, a new database is created in a
system specific temporary folder.
}
\section{Slots}{

\describe{
\item{\code{name}}{The name of the database file.}

\item{\code{file}}{The absolute file path of the database file.}

\item{\code{handle}}{An external pointer to the sqlite3 database connection.}
}}

