% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_db.R
\name{db.close}
\alias{db.close}
\title{Generic function to close a database connection}
\usage{
db.close(con, showWarnings = TRUE)
}
\arguments{
\item{con}{database connection to be closed}

\item{showWarnings}{logical: report possible issues with connection?}
}
\value{
`TRUE`, invisibly (see [DBI::dbDisconnect()])
}
\description{
Close a previously opened connection to a database.
}
\examples{
\dontrun{
db.close(con)
}
}
\author{
Rob Kooper
}
