% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\docType{data}
\name{DB.connections}
\alias{DB.connections}
\title{An environment to store database connections}
\format{
An object of class \code{environment} of length 0.
}
\usage{
DB.connections
}
\description{
An environment to store database connections
}
\details{
Database connections are registered under their
file name. Temporary on-disk and in-memory databases are
never registered in this environment.
}
\keyword{datasets}
