% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-vtable.R
\name{db.create_module}
\alias{db.create_module}
\title{Register a virtual table module}
\usage{
db.create_module(module)
}
\arguments{
\item{module}{The virtual table module. An S4 object of class "virtualtable".}
}
\value{
None
}
\description{
Register a virtual table module
}
