% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-vtable.R
\name{db.delete_module}
\alias{db.delete_module}
\title{Unregister a virtual table module}
\usage{
db.delete_module(module)
}
\arguments{
\item{module}{The virtual table module. An S4 object of class "virtualtable".}
}
\value{
None
}
\description{
Unregister a virtual table module
}
