% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-eval.R
\name{db.exec}
\alias{db.exec}
\title{Execute a SQL script}
\usage{
db.exec(db, script)
}
\arguments{
\item{db}{The database connection. An S4 object of class "database".}

\item{sql}{The SQL script to evaluate.}
}
\description{
Run one or more semi-colon terminated SQL statements
}
