% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-info.R
\name{db.exists}
\alias{db.exists}
\title{List the tables in a database}
\usage{
db.exists(db, name)
}
\arguments{
\item{db}{The database connection. S4 object of class "database".}

\item{name}{The name of the table, view, or index in the database.}
}
\value{
\code{TRUE} if the item exists. Otherwise \code{FALSE}.
}
\description{
List the tables in a database
}
