% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-eval.R
\name{db.fetch}
\alias{db.fetch}
\title{Retrieve a row from the result set of a query}
\usage{
db.fetch(cur, df = FALSE)
}
\arguments{
\item{cur}{The cursor returned by \code{db.eval}. An S4 object of class
"cursor".}

\item{df}{A boolean. Should the result be returned as a data.frame?}
}
\value{
A named list (df=FALSE) or data.frame (df=TRUE) containing a
row from the result set.
}
\description{
Retrieve a row from the result set of a query
}
