% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-eval.R
\name{db.fetchall}
\alias{db.fetchall}
\title{Retrieve all rows from the result set of a query}
\usage{
db.fetchall(cur, df = FALSE)
}
\arguments{
\item{cur}{The cursor returned by \code{db.eval}. An S4 object of class
"cursor".}

\item{df}{A boolean. Should the result be returned as a data.frame?}
}
\value{
A matrix (df=FALSE) or data.frame (df=TRUE) containing the rows
from the result set. If \code{df=FALSE} subsets of the matrix will
return lists rather than atomic vectors.
}
\description{
Retrieve all rows from the result set of a query
}
