% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-info.R
\name{db.fields}
\alias{db.fields}
\title{List the fields of a database table}
\usage{
db.fields(db, table)
}
\arguments{
\item{db}{The database connection. S4 object of class "database".}

\item{table}{The name of a table in the database.}
}
\value{
A data.frame listing the name and storage mode of each field
in the table.
}
\description{
List the fields of a database table
}
