% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-function.R
\name{db.function}
\alias{db.function}
\title{Register functions with a database connection}
\usage{
db.function(db, NAME, FUN)
}
\arguments{
\item{db}{The database connection. S4 object of class "database".}

\item{NAME}{The name to register the function under.}

\item{FUN}{The function to register.}
}
\value{
None.
}
\description{
Register functions with a database connection
}
\details{
Functions registered with a database connection can be referred to
by NAME in SQL statements. To unregister a function NAME, pass NULL as the
FUN argument. See examples.
}
\note{
A current limitation is that arguments to functions inside SQL
statements can only be referred to by position. In other words, constructs
like argname=argvalue will not work.
}
\examples{
db = db.open()
db.function(db, "multiply", `*`)
db.eval(db, "CREATE TABLE t(a REAL, b REAL)")
db.eval(db, "INSERT INTO t VALUES (?,?)", matrix(rnorm(200), 100, 2))
cursor = db.eval(db, "SELECT multiply(a, b) FROM t")
db.fetchall(cursor, TRUE)
db.function(db, "multiply", NULL)
\dontrun{
db.eval(db, "SELECT multiply(a, b) FROM t") # no such function: multiply
}
}
