% Generated by roxygen2 (4.0.1): do not edit by hand
\name{db.getShowQueries}
\alias{db.getShowQueries}
\title{db.getShowQueries}
\usage{
db.getShowQueries()
}
\value{
will return TRUE if queries are shown
}
\description{
Returns if the queries should be shown that are being executed
}
\author{
Rob Kooper
}

