% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-httpd.R
\name{db.help}
\alias{db.help}
\title{Open the help documentation for a database}
\usage{
db.help(db)
}
\arguments{
\item{db}{An S4 object of class "database".}
}
\description{
Open the help documentation for a database
}
\seealso{
\code{\link{db.httpd}}
}
