% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-json.R
\name{db.json}
\alias{db.json}
\title{Dump the contents of a table or database into a JSON array}
\usage{
db.json(db, table, file = NULL)
}
\arguments{
\item{db}{The database connection.}

\item{table}{The table name.}

\item{file}{A file to dump the contents to. If NULL the dump is
returned as a character string.}
}
\description{
Dump the contents of a table or database into a JSON array
}
\details{
If \code{table} is missing the entire database is dumped
into an array of JSON objects. Each object has two fields: a "table"
field with the name of the table and a "data" field with the table
data. The "data" is an array of JSON objects. Each object represents
one row of table data and the JSON object holds a field for each table
column. If \code{table} is not missing then only the contents of that
table are dumped and the return value consists only of the "data" array.
}
\note{
This does require that SQLite was compiled with the option
-DSQLITE_ENABLE_JSON1.
}
