% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-eval.R
\name{db.lapply}
\alias{db.lapply}
\title{Apply a function to rows in a result set}
\usage{
db.lapply(db, stmt, params, FUN, ...)
}
\arguments{
\item{db}{The database connection. An S4 object of class "database".}

\item{stmt}{The SQL statement to evaluate.}

\item{params}{A matrix, data.frame, or list of lists with parameters.}

\item{FUN}{The function to apply.}

\item{...}{Additional arguments to FUN.}
}
\value{
A list, each element of which is the result of applying FUN to
a row in the result set.
}
\description{
Apply a function to rows in a result set
}
\details{
Function FUN is called on each row returned by the query stmt.
The first argument to FUN will be a named list containing the fields
for a row returned by the stmt. Subsequent arguments to FUN are passed
via ...
}
\examples{
db = db.open()
db.eval(db, "CREATE TABLE t(a REAL, b REAL)")
db.eval(db, "INSERT INTO t VALUES(?,?)", matrix(runif(200), 100, 2))
db.lapply(db, "SELECT * FROM t", list(list()), function(r) sum(unlist(r)))
}
\seealso{
\code{\link{db.eval}}
}
