% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-eval.R
\name{db.param}
\alias{db.param}
\title{Convert input parameters to expected format (list of lists)}
\usage{
db.param(params)
}
\description{
This is an internal function and not meant to be
called outside the db package.
}
