% Generated by roxygen2 (4.0.2): do not edit by hand
\name{db.query.file.path}
\alias{db.query.file.path}
\alias{query.file.path}
\title{Get file path given id and machine}
\usage{
query.file.path(input.id, host_name, con)
}
\arguments{
\item{host_name}{}

\item{con}{: database connection}

\item{input_id}{}
}
\description{
Get file path given id and machine
}
\author{
Betsy Cowdery
}

