% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_db.R
\name{db.query}
\alias{db.query}
\title{Generic function to query database}
\usage{
db.query(query, con = NULL, params = NULL)
}
\arguments{
\item{query}{SQL query string}

\item{con}{Database connection object}

\item{params}{Named list of database connection parameters. See
`params` argument to [db.open()].}
}
\value{
data frame with query results
}
\description{
Given a connection and a query, will return a query as a data frame. Either con or params need
to be specified. If both are specified it will use con.
}
\examples{
\dontrun{
db.query("SELECT count(id) FROM traits;", params = settings$database$bety)
}
}
\author{
Rob Kooper, Alexey Shiklomanov
}
