% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-io.R
\name{db.register}
\alias{db.register}
\title{Register a database connection}
\usage{
db.register(db)
}
\arguments{
\item{db}{An S4 object of class "database".}
}
\description{
This is done automatically by \code{\link{db.open}}.
Databases are registered under their file name.
}
