% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-vtable.R
\name{db.register_shadowname}
\alias{db.register_shadowname}
\title{Register a the name of shadow table for a virtual table}
\usage{
db.register_shadowname(db, name)
}
\arguments{
\item{db}{The database connection. An S4 object of class "database".}

\item{name}{The shadow name.}
}
\value{
None
}
\description{
For details on shadow tables, including how SQLite recognizes a shadow
table, go to \url{https://sqlite.org/vtab.html#the_xshadowname_method}.
}
\details{
When a shadow name is registered with a database connection
the contents of the table bearing the shadow name cannot be altered
except by SQL that originates from one of the virtual table methods.
}
