% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-httpd.R
\name{db.reply}
\alias{db.reply}
\title{Add text to the HTTP response}
\usage{
db.reply(resp, text)
}
\arguments{
\item{resp}{The HTTP response being formed}

\item{reply}{Text to append to the HTTP response}
}
\description{
Add text to the HTTP response
}
