% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-httpd.R
\name{db.reply_file}
\alias{db.reply_file}
\title{Make the contents of a file the HTTP response}
\usage{
db.reply_file(resp, filename)
}
\arguments{
\item{resp}{The HTTP response being formed}

\item{filename}{Name of the file whose content forms the response}
}
\description{
Make the contents of a file the HTTP response
}
