% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-httpd.R
\name{db.reply_raw}
\alias{db.reply_raw}
\title{Return a raw vector as the HTTP response}
\usage{
db.reply_raw(resp, raw)
}
\arguments{
\item{resp}{The HTTP response being formed}

\item{raw}{Binary data that forms the response}
}
\description{
Return a raw vector as the HTTP response
}
