% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-info.R
\name{db.schema}
\alias{db.schema}
\title{List the schema of database tables}
\usage{
db.schema(db, table)
}
\arguments{
\item{db}{The database connection. S4 object of class "database".}

\item{table}{A character vector of table names in the database.}
}
\value{
A data.frame listing the name and SQL schema for each table.
}
\description{
List the schema of database tables
}
