% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{db.showQueries}
\alias{db.showQueries}
\title{db.showQueries}
\usage{
db.showQueries(show)
}
\arguments{
\item{show}{set to TRUE to show the queries, FALSE by default}
}
\description{
Sets if the queries should be shown that are being executed
}
\details{
Useful to print queries when debuging SQL statements
}
\author{
Rob Kooper
}

