% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-sqlar.R
\name{db.sqlar}
\alias{db.sqlar}
\title{Create a SQLite archive}
\usage{
db.sqlar(db, path)
}
\arguments{
\item{db}{The database connection. S4 object of class "database".}

\item{path}{The path to the top-level directory of the archive.}
}
\value{
None.
}
\description{
For additional details goto \url{https://sqlite.org/sqlar.html}.
}
\details{
This will create a table named "sqlar" in the database. The
table will hold the gzip compressed content of all files and
directories under \code{path}.
}
\note{
The \code{\link{sqlar_compress}} function must be registered
with the database connection for this function to work. This is done
automatically with \code{\link{db.open}}
}
\seealso{
\code{\link{db.sqlar.skeleton}}
}
