% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-sqlar.R
\name{db.sqlar.skeleton}
\alias{db.sqlar.skeleton}
\title{Initialize the SQLite archive schema}
\usage{
db.sqlar.skeleton(db)
}
\arguments{
\item{db}{The database connection. S4 object of class "database".}
}
\value{
None.
}
\description{
Initialize the SQLite archive schema
}
\details{
This has a side effect of creating a table named "sqlar"
in the database file pointed to by \code{db}. The sqlar table has
the following schema:
\preformatted{
    CREATE TABLE sqlar(
        name  TEXT PRIMARY KEY, -- name of the file
        mode  INTEGER,          -- access permissions
        mtime INTEGER,          -- last modification time
        sz    INTEGER,          -- original file size
        data  BLOB              -- compressed content
    );
}
}
