% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-sqlar.R
\name{db.sqlar.update}
\alias{db.sqlar.update}
\title{Update the content of a SQLite archive}
\usage{
db.sqlar.update(db, path)
}
\arguments{
\item{db}{The database connection. S4 object of class "database".}

\item{path}{The path to the top-level directory of the archive.}
}
\value{
None.
}
\description{
Update the content of a SQLite archive
}
\details{
This will update the sqlar table in the database to reflect
any changes made to the files under \code{path} (including file
additions and deletions and changed file permissions).
}
\note{
The \code{\link{sqlar_compress}} function must be registered
with the database connection for this function to work. This is done
automatically with \code{\link{db.open}}
}
