% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-sqlar.R
\name{db.sqlar_root}
\alias{db.sqlar_root}
\title{Return the root directory of a SQLite archive}
\usage{
db.sqlar_root(db, name)
}
\arguments{
\item{db}{The database connection. S4 object of class "database".}

\item{name}{The name of the SQLite archive table.}
}
\description{
Return the root directory of a SQLite archive
}
