% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-info.R
\name{db.tables}
\alias{db.tables}
\title{List the tables in a database}
\usage{
db.tables(db)
}
\arguments{
\item{db}{The database connection. S4 object of class "database".}
}
\value{
A character vector of table names in the database.
}
\description{
List the tables in a database
}
