% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-io.R
\name{db.unregister}
\alias{db.unregister}
\title{Unregister a database connection}
\usage{
db.unregister(db)
}
\arguments{
\item{db}{An S4 object of class "database".}
}
\description{
This is done automatically by \code{\link{db.close}}
}
