% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-vtable.R
\name{db.unregister_shadowname}
\alias{db.unregister_shadowname}
\title{Unregister a shadow name}
\usage{
db.unregister_shadowname(db, name)
}
\arguments{
\item{db}{The database connection. An S4 object of class "database".}

\item{name}{The shadow name.}
}
\value{
None
}
\description{
Unregister a shadow name
}
