% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-sqlar.R
\name{db.unsqlar}
\alias{db.unsqlar}
\title{Unpack a SQLite archive.}
\usage{
db.unsqlar(db, name, path, files)
}
\arguments{
\item{db}{The database connection. S4 object of class "database".}

\item{name}{The name of the SQLite archive table.}

\item{path}{The path to unpack the archive under.}

\item{files}{The files in the archive to extract. These need not be
complete names. For example, "file1" will be matched using wildcards
on either end. So if "file1" is a directory all files under it will be
extracted.}
}
\value{
None.
}
\description{
Unpack a SQLite archive.
}
\details{
This will read the sqlar table in the database and write out
its content into a filesystem hierarchy under \code{path}.
}
\note{
The \code{\link{sqlar_uncompress}} function must be registered
with the database connection for this function to work. This is done
automatically with \code{\link{db.open}}
}
