% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfiles.R
\name{dbfile.check}
\alias{dbfile.check}
\title{Check for a file in the dbfiles tables}
\usage{
<<<<<<< HEAD
dbfile.check(type, id, con, hostname = fqdn(), machine.check = TRUE)
}
\arguments{
=======
dbfile.input.check(siteid, startdate = NULL, enddate = NULL, mimetype,
  formatname, parentid = NA, con, hostname = fqdn(), exact.dates = FALSE)

dbfile.posterior.check(pft, mimetype, formatname, con, hostname = fqdn())

dbfile.check(type, id, con, hostname = fqdn())
}
\arguments{
\item{siteid}{the id of the site that this data is applicable to}

\item{startdate}{the start date of the data stored in the file}

\item{enddate}{the end date of the data stored in the file}

\item{mimetype}{the mime-type of the file}

\item{formatname}{the name of the format to distinguish between simmilair mime-types}

\item{con}{database connection object}

\item{hostname}{the name of the host where the file is stored, this will default to the name of the current machine}

\item{exact.dates}{setting to include start and end date in input query}

\item{pft}{the name of the pft that this data is applicable to}

>>>>>>> dacd76a384839ef9b7c0f2ecf88b3bdd18651e28
\item{type}{the type of dbfile (Input, Posterior)}

\item{id}{the id of container type}

\item{con}{database connection object}

\item{hostname}{the name of the host where the file is stored, this will default to the name of the current machine}

\item{machine.check}{setting to check for file on named host, otherwise will check for any file given container id}
}
\value{
data.frame with the id, filename and pathname of all the files that are associated
}
\description{
Function to check to see if a file exists in the dbfiles table
}
\details{
This will check the dbfiles and machines to see if the file exists
}
\examples{
\dontrun{
  dbfile.check('Input', 7, dbcon)
}
}
\author{
<<<<<<< HEAD
=======
Rob Kooper, Tony Gardella

Rob Kooper

>>>>>>> dacd76a384839ef9b7c0f2ecf88b3bdd18651e28
Rob Kooper
}

