% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfiles.R
\name{dbfile.file}
\alias{dbfile.file}
\alias{dbfile.id}
\title{Convert between file paths and ids}
\usage{
dbfile.file(type, id, con, hostname = PEcAn.remote::fqdn())

dbfile.id(type, file, con, hostname = PEcAn.remote::fqdn())
}
\arguments{
\item{type}{the type of dbfile (Input, Posterior)}

\item{id}{the id of container type}

\item{con}{database connection object}

\item{hostname}{the name of the host where the file is stored, this will default to the name of the current machine}

\item{file}{the full pathname to the file}
}
\value{
filename on host, or NA if none found
}
\description{
These functions check the dbfiles and machines tables to see if the file
exists, and return the container_id (\code{dbfile.id}) or full filename
with path (\code{dbfile.file}) to the first one found.
If none is found, both will return NA.
}
\section{Functions}{
\itemize{
\item \code{dbfile.file}: Return full path to file from the dbfiles table

\item \code{dbfile.id}: Return id to container type given a filename
}}

\examples{
\dontrun{
  dbfile.file('Input', 7, dbcon)
}
\dontrun{
  dbfile.id('Model', '/usr/local/bin/sipnet', dbcon)
}
}
\author{
Rob Kooper

Rob Kooper
}
