% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfiles.R
\name{dbfile.format.variable.insert}
\alias{dbfile.format.variable.insert}
\title{Function to Insert Records into the Formats and Variables Tables}
\usage{
dbfile.format.variable.insert(format_name, header, skip, mimetype_id,
  format_notes = "", con, format_variables = FALSE, description = NA,
  units = NA, var_notes = NA, var_name = NA, max = NA, min = NA,
  standard_name = NA, label = NA, type = NA)
}
\arguments{
\item{format_name}{Format Name}

\item{header}{TRUE/FALSE Boolean describing whether or not the file contains a header}

\item{skip}{The number of lines in the header}

\item{mimetype_id}{The mimetype id of the mimetype that is associated with the format}

\item{format_notes}{Any useful notes that describe the format being created}

\item{con}{Connection object to connect to bety}

\item{format_variables}{A TRUE/FALSE Boolean used to toggle a format-variables db query}

\item{description}{variable descripthon}

\item{units}{variable units}

\item{var_notes}{notes to further describe the variable}

\item{var_name}{variable name}

\item{max}{maximum possible value}

\item{min}{minimum possible value}

\item{standard_name}{PEcAn specific name?}

\item{label}{?}

\item{type}{?}
}
\value{
formatid
}
\description{
Function to Insert Records into the Formats and Variables Tables
}
\examples{
\dontrun{
bety <- betyConnect()
  dbfile.format.variable.insert('Test_Format', header = TRUE, skip = 2, mimetype_id = 1060, con = bety$con, format_variables = FALSE)
}
}
\author{
Liam Burke
}
