% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfiles.R
\name{dbfile.input.check}
\alias{dbfile.input.check}
\title{Check for a file in the input/dbfiles tables}
\usage{
dbfile.input.check(siteid, startdate = NULL, enddate = NULL, mimetype,
  formatname, parentid = NA, con, hostname = PEcAn.remote::fqdn(),
  exact.dates = FALSE, pattern = NULL)
}
\arguments{
\item{siteid}{the id of the site that this data is applicable to}

\item{startdate}{the start date of the data stored in the file}

\item{enddate}{the end date of the data stored in the file}

\item{mimetype}{the mime-type of the file}

\item{formatname}{the name of the format to distinguish between simmilair mime-types}

\item{parentid}{the id of the parent of the input}

\item{con}{database connection object}

\item{hostname}{the name of the host where the file is stored, this will default to the name of the current machine}

\item{exact.dates}{setting to include start and end date in input query}

\item{pattern}{text to seach for in the file name (default NULL = no check).}
}
\value{
data.frame with the id, filename and pathname of the input that is requested
}
\description{
Function to check to see if a file exists in the dbfiles table as an input
}
\details{
This will check the dbfiles, inputs, machines and formats tables to see if the
file exists
}
\examples{
\dontrun{
  dbfile.input.check(siteid, startdate, enddate, 'application/x-RData', 'traits', dbcon)
}
}
\author{
Rob Kooper, Tony Gardella
}
