% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfiles.R
\name{dbfile.insert}
\alias{dbfile.insert}
\title{Insert file into tables}
\usage{
dbfile.insert(in.path, in.prefix, type, id, con, reuse = TRUE,
  hostname = PEcAn.remote::fqdn())
}
\arguments{
\item{in.path}{Path to file directory}

\item{in.prefix}{Filename prefix (not including directory)}

\item{con}{database connection object}

\item{hostname}{the name of the host where the file is stored, this will default to the name of the current machine}

\item{params}{database connection information}
}
\value{
id of the file that is written
}
\description{
Function to insert a file into the dbfiles table
}
\details{
This will write into the dbfiles and machines the required data to store the file
}
\examples{
\dontrun{
  dbfile.insert('somefile.txt', 'Input', 7, dbcon)
}
}
\author{
Rob Kooper, Ryan Kelly
}
