% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfiles.R
\name{dbfile.posterior.check}
\alias{dbfile.posterior.check}
\title{Check for a file in the input/dbfiles tables}
\usage{
dbfile.posterior.check(pft, mimetype, formatname, con,
  hostname = PEcAn.remote::fqdn())
}
\arguments{
\item{pft}{the name of the pft that this data is applicable to}

\item{mimetype}{the mime-type of the file}

\item{formatname}{the name of the format to distinguish between simmilair mime-types}

\item{con}{database connection object}

\item{hostname}{the name of the host where the file is stored, this will default to the name of the current machine}
}
\value{
data.frame with the id, filename and pathname of the posterior that is requested
}
\description{
Function to check to see if a file exists in the dbfiles table as an input
}
\details{
This will check the dbfiles, inputs, machines and formats tables to see if the
file exists
}
\examples{
\dontrun{
  dbfile.posterior.check(pft, 'application/x-RData', 'traits', dbcon)
}
}
\author{
Rob Kooper
}
