% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.dplyr.R
\name{fancy_scientific}
\alias{fancy_scientific}
\title{Convert number to scientific notation pretty expression}
\usage{
fancy_scientific(l)
}
\arguments{
\item{l}{Number to convert to scientific notation}
}
\description{
Convert number to scientific notation pretty expression
}
