% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{fn_add_new_columns}
\alias{fn_add_new_columns}
\title{Add new columns into an existing data table in the database}
\usage{
fn_add_new_columns(df, database, table_name, verbose = TRUE)
}
\arguments{
\item{df}{data frame representing a data table, e.g. phenotypes, environments or genotypes table}

\item{database}{an open SQLite database connection}

\item{table_name}{name of the data table represented by df}

\item{verbose}{Show messages? (Default=TRUE)}
}
\value{
\itemize{
\item Ok: database - the same open SQLite database connection from the input
\item Err: dbError
}
}
\description{
Add new columns into an existing data table in the database
}
\examples{
list_fnames_tables = fn_simulate_tables(
     n_entries=50,
     n_dates=3,
     n_sites=3,
     n_treatments=3,
     n_loci=10e3,
     save_data_tables=TRUE)$list_fnames_tables
df = utils::read.delim(list_fnames_tables$fname_phenotypes, header=TRUE)
database = DBI::dbConnect(drv=RSQLite::SQLite(), dbname="test.sqlite")
### Prepare the tables
list_df_data_and_base_tables = fn_prepare_data_table_and_extract_base_tables(df=df, 
     database=database, table_name="phenotypes", verbose=TRUE)
df = list_df_data_and_base_tables$df_possibly_modified
df_first_half = droplevels(df[, 1:floor(ncol(df)/2)])
### Import the first half of the data table into the database
DBI::dbWriteTable(conn=database, name="phenotypes", value=df_first_half)
### Add new columns
database = fn_add_new_columns(df=df, database=database, table_name="phenotypes", verbose=TRUE)
DBI::dbDisconnect(database)
unlink("test.sqlite")
}
