% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{fn_add_POSIX_time}
\alias{fn_add_POSIX_time}
\title{Add POSIX time column into a data table}
\usage{
fn_add_POSIX_time(
  df,
  database,
  table_name,
  bool_add_FVI_year_season = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{df}{data frame representing a base (entries, dates, sites, treatments, traits, abiotics,
and loci) or data (phenotypes, environments and genotypes) table}

\item{database}{an open SQLite database connection}

\item{table_name}{name of the base or data table represented by df}

\item{bool_add_FVI_year_season}{Add the corresponding forage value index (FVI) seasons and year?
(Default=TRUE)}

\item{verbose}{Show messages? (Default=TRUE)}
}
\value{
\itemize{
\item Ok:
df: data frame
\item Err: dbError
}
}
\description{
Add POSIX time column into a data table
}
\examples{
list_fnames_tables = fn_simulate_tables(
     n_entries=50,
     n_dates=3,
     n_sites=3,
     n_treatments=3,
     n_loci=10e3,
     save_data_tables=TRUE)$list_fnames_tables
df = utils::read.delim(list_fnames_tables$fname_phenotypes, header=TRUE)
database = DBI::dbConnect(drv=RSQLite::SQLite(), dbname="test.sqlite")
df = fn_add_POSIX_time(df=df, database=database, table_name="phenotypes", 
     bool_add_FVI_year_season=TRUE, verbose=TRUE)
DBI::dbDisconnect(database)
unlink("test.sqlite")
}
