% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{fn_check_import_inputs}
\alias{fn_check_import_inputs}
\title{Check the validity of importation function inputs}
\usage{
fn_check_import_inputs(
  df,
  database,
  table_name,
  check_UIDs = FALSE,
  check_if_table_exists = FALSE
)
}
\arguments{
\item{df}{data frame representing a base (entries, dates, sites, treatments, traits, abiotics,
and loci) or data (phenotypes, environments and genotypes) table}

\item{database}{an open SQLite database connection}

\item{table_name}{name of the base or data table represented by df}

\item{check_UIDs}{Does the table-specific UID column exist in df? (Default=FALSE)}

\item{check_if_table_exists}{Does the table exist in the database? (Default=FALSE)}
}
\value{
\itemize{
\item Ok:
NULL
\item Err: dbError
}
}
\description{
Check the validity of importation function inputs
}
\examples{
list_fnames_tables = fn_simulate_tables(
     n_entries=50,
     n_dates=3,
     n_sites=3,
     n_treatments=3,
     n_loci=10e3,
     save_data_tables=TRUE)$list_fnames_tables
df = utils::read.delim(list_fnames_tables$fname_phenotypes, header=TRUE)
database = DBI::dbConnect(drv=RSQLite::SQLite(), dbname="test.sqlite")
null_error = fn_check_import_inputs(df=df, database=database, table_name="phenotypes", 
 check_UIDs=FALSE)
non_null_error = fn_check_import_inputs(df=df[, -1:-5], database=database, 
     table_name="phenotypes", check_UIDs=FALSE)
DBI::dbDisconnect(database)
unlink("test.sqlite")
}
