% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{fn_define_hash_and_UID_prefix}
\alias{fn_define_hash_and_UID_prefix}
\title{Define a consistent prefix to hash and UID column names for each table type}
\usage{
fn_define_hash_and_UID_prefix(table_name)
}
\arguments{
\item{table_name}{name of the base (entries, dates, sites, treatments, traits, abiotics,
and loci) or data (phenotypes, environments and genotypes) table}
}
\value{
\itemize{
\item Ok:
prefix_of_HASH_and_UID_columns: character string
\item Err: dbError
}
}
\description{
Define a consistent prefix to hash and UID column names for each table type
}
\examples{
prefix_of_HASH_and_UID_columns = fn_define_hash_and_UID_prefix(table_name="phenotypes")
}
