% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{fn_deserialise_genotype_data}
\alias{fn_deserialise_genotype_data}
\title{Deserialise the genotype data from the genotypes table in the database}
\usage{
fn_deserialise_genotype_data(database, df_genotypes, verbose = TRUE)
}
\arguments{
\item{database}{an open SQLite database connection}

\item{df_genotypes}{data frame containing the genotype data extracted from the database.
The first column refers to the ENTRY_UIDs, and the second column refer to the BLOBs of
serialiased/binary genotype data.}

\item{verbose}{Show messages? (Default=TRUE)}
}
\value{
\itemize{
\item Ok:
df_allele_frequency_table: data frame corresponding to an allele frequency table, where
the first 3 columns refer to the CHROMOSOME, POSITION_PER_CHROMOSOME, and ALLELE,
followed by the allele frequencies per sample, entry or pool with their corresponding
entry names as the column names.
\item Err: dbError
}
}
\description{
Deserialise the genotype data from the genotypes table in the database
}
\examples{
list_fnames_tables = fn_simulate_tables(
     n_entries=50,
     n_dates=3,
     n_sites=3,
     n_treatments=3,
     n_loci=10e3,
     save_data_tables=TRUE)$list_fnames_tables
fname_db = "test.sqlite"
df_allele_frequency_table = utils::read.delim(
         list_fnames_tables$fname_genotypes, 
         header=TRUE, check.names=FALSE)
list_df_data_tables = list(
    df_phenotypes=utils::read.delim(list_fnames_tables$fname_phenotypes, header=TRUE),
    df_environments=utils::read.delim(list_fnames_tables$fname_environments, header=TRUE),
    df_genotypes=df_allele_frequency_table)
fn_initialise_db(fname_db=fname_db, list_df_data_tables=list_df_data_tables, verbose=TRUE)
database = DBI::dbConnect(drv=RSQLite::SQLite(), dbname=fname_db)
df_genotypes = DBI::dbGetQuery(conn=database, statement="SELECT * FROM genotypes")
df_allele_frequency_table = fn_deserialise_genotype_data(database=database, df_genotypes=df_genotypes)
DBI::dbDisconnect(database)
unlink("test.sqlite")
}
