% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fn_export_phenotypes_and_genotypes_data_from_database}
\alias{fn_export_phenotypes_and_genotypes_data_from_database}
\title{Export phenotype and genotype data from the database using vectors of}
\usage{
fn_export_phenotypes_and_genotypes_data_from_database(
  fname_db,
  vec_ENTRY = c("*"),
  vec_REPLICATION = c("*"),
  vec_TREATMENT = c("*"),
  vec_SITE = c("*"),
  vec_YEAR = c("*"),
  vec_MONTH = c("*"),
  vec_DAY = c("*"),
  vec_trait_names = c("*"),
  fname_basename_out = NULL,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{fname_db}{name of the SQLite database file.}

\item{vec_ENTRY}{vector of entry names to extract. (Default=c("*"))}

\item{vec_REPLICATION}{vector of replications to extract. (Default=c("*"))}

\item{vec_TREATMENT}{vector of treatments to extract. (Default=c("*"))}

\item{vec_SITE}{vector of sites to extract. (Default=c("*"))}

\item{vec_YEAR}{vector of years to extract. (Default=c("*"))}

\item{vec_MONTH}{vector of months to extract. (Default=c("*"))}

\item{vec_DAY}{vector of days to extract. (Default=c("*"))}

\item{vec_trait_names}{vector of trait names to include in the phenotypes table. (Default=c("*"))}

\item{fname_basename_out}{base name of the output/queried phenotype and genotype files. (Default=NULL)}

\item{overwrite}{Overwrite the queried/output phenotypes and genotypes files? (Default=FALSE)}

\item{verbose}{Show messages? (Default=TRUE)}
}
\value{
\itemize{
\item Ok:
$fname_phenotypes_tsv: name of the output/queried phenotypes table file.
This file is tab-delimited.
$fname_genotypes_tsv: name of the output/queried allele frequency table file.
This is a tab-delimited file with a header line and the first 3 columns refer to the
chromosome (chr), position (pos), and allele (allele), with subsequent columns referring
to the allele frequencies of a sample.
\item Err: dbError
}
}
\description{
Export phenotype and genotype data from the database using vectors of
}
\examples{
fname_xlsx = fn_simulate_tables(
    n_entries=50,
    n_dates=3,
    n_sites=3,
    n_treatments=3,
    n_loci=10e3,
    save_data_tables=TRUE)$list_fnames_tables$fname_data_tables
fname_db = "test.sqlite"
fn_create_database_from_xlsx_or_tsv(fname_db=fname_db, fname_xlsx=fname_xlsx, overwrite=TRUE)
database = DBI::dbConnect(drv=RSQLite::SQLite(), dbname=fname_db)
vec_entry_names = sample(DBI::dbGetQuery(conn=database, 
     statement="SELECT ENTRY FROM entries")[, 1], size=3)
DBI::dbDisconnect(conn=database)
list_fnames_out = fn_export_phenotypes_and_genotypes_data_from_database(
     fname_db=fname_db, 
     vec_ENTRY=vec_entry_names)
unlink(list_fnames_out$fname_phenotypes)
unlink(list_fnames_out$fname_genotypes)
}
