% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{fn_initialise_db}
\alias{fn_initialise_db}
\title{Initialise the database}
\usage{
fn_initialise_db(fname_db, list_df_data_tables, verbose = TRUE)
}
\arguments{
\item{fname_db}{name of the SQLite database file}

\item{list_df_data_tables}{list containing the 3 data tables as data frames,
i.e. 'df_phenotypes', 'df_environments' and 'df_genotypes' data tables}

\item{verbose}{Show messages? (Default=TRUE)}
}
\value{
\itemize{
\item Ok: 0
\item Err: dbError
}
}
\description{
Initialise the database
}
\examples{
fname_data_tables = fn_simulate_tables(
     n_entries=50,
     n_dates=3,
     n_sites=3,
     n_treatments=3,
     n_loci=10e3,
     save_data_tables=TRUE)$list_fnames_tables$fname_data_tables
list_df_data_tables = list()
for (table_name in c("phenotypes", "environments", "genotypes")) {
    eval(parse(text=paste0("list_df_data_tables$df_", table_name, 
         " = as.data.frame(readxl::read_excel(path=fname_data_tables, sheet=table_name))")))
}
fn_initialise_db(fname_db="test.sqlite", list_df_data_tables=list_df_data_tables)
}
