% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{fn_prepare_data_table_and_extract_base_tables}
\alias{fn_prepare_data_table_and_extract_base_tables}
\title{Prepare a data table and extract based tables from it}
\usage{
fn_prepare_data_table_and_extract_base_tables(
  df,
  database,
  table_name,
  verbose = TRUE
)
}
\arguments{
\item{df}{data frame representing a data table, e.g. phenotypes, environments or genotypes table}

\item{database}{an open SQLite database connection}

\item{table_name}{name of the base or data table represented by df}

\item{verbose}{Show messages? (Default=TRUE)}
}
\value{
\itemize{
\item Ok:
$df_possibly_modified: data table with column names converted to uppercase, hash and UID
columns added, and duplicate rows and column ommitted
$df_entries: "entries" base table or data frame
$df_dates: "dates" base table or data frame
$df_sites: "sites" base table or data frame
$df_treatments: "treatments" base table or data frame
$df_traits: "traits" base table or data frame
$df_abiotics: "abiotics" base table or data frame
$df_loci: "loci" base table or data frame
\item Err: dbError
}
}
\description{
Prepare a data table and extract based tables from it
}
\examples{
list_fnames_tables = fn_simulate_tables(
     n_entries=50,
     n_dates=3,
     n_sites=3,
     n_treatments=3,
     n_loci=10e3,
     save_data_tables=TRUE)$list_fnames_tables
df = utils::read.delim(list_fnames_tables$fname_phenotypes, header=TRUE)
database = DBI::dbConnect(drv=RSQLite::SQLite(), dbname="test.sqlite")
list_df_data_and_base_tables = fn_prepare_data_table_and_extract_base_tables(df=df, 
 database=database, table_name="phenotypes", verbose=TRUE)
DBI::dbDisconnect(database)
unlink("test.sqlite")
}
