% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{fn_remove_quotes_and_newline_characters_in_data}
\alias{fn_remove_quotes_and_newline_characters_in_data}
\title{Remove quotes and newline characters ("\\n" in Linux and "\\r\\n" in Windows)}
\usage{
fn_remove_quotes_and_newline_characters_in_data(df, verbose = TRUE)
}
\arguments{
\item{df}{data frame representing a base (entries, dates, sites, treatments, traits, abiotics,
and loci) or data (phenotypes, environments and genotypes) table}

\item{verbose}{Show messages? (Default=TRUE)}
}
\value{
\itemize{
\item Ok:
df: data frame
\item Err: dbError
}
}
\description{
Remove quotes and newline characters ("\\n" in Linux and "\\r\\n" in Windows)
}
\examples{
list_fnames_tables = fn_simulate_tables(
     n_entries=50,
     n_dates=3,
     n_sites=3,
     n_treatments=3,
     n_loci=10e3,
     save_data_tables=TRUE)$list_fnames_tables
df = utils::read.delim(list_fnames_tables$fname_phenotypes, header=TRUE)
df[1,1] = paste0("'", df[1,1], "'")
df[2,2] = paste0('"', df[2,2], '"')
df[3,3] = paste0(df[3,3], "\n")
df[4,4] = paste0(df[4,4], "\r\n")
df = fn_remove_quotes_and_newline_characters_in_data(df=df, verbose=TRUE)
}
