% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{fn_set_classification_of_columns}
\alias{fn_set_classification_of_columns}
\title{Classify the columns of the incoming base or data table to check for intersection/s with existing
data and base data tables in the database}
\usage{
fn_set_classification_of_columns(df, database, table_name, verbose = TRUE)
}
\arguments{
\item{df}{data frame representing a base (entries, dates, sites, treatments, traits, abiotics,
and loci) or data (phenotypes, environments and genotypes) table}

\item{database}{an open SQLite database connection}

\item{table_name}{name of the base or data table represented by df}

\item{verbose}{Show messages? (Default=TRUE)}
}
\value{
\itemize{
\item Ok:
$n_existing_columns: total number of columns in the existing table
$n_incoming_columns: total number of columns in the incoming table
$n_intersecting_columns: number of intersecting columns in the existing and incoming table
$n_columns_exclusive_to_existing_table: number of columns exclusive to the existing table
$n_columns_exclusive_to_incoming_table: number of columns exclusive to the incoming table
$vec_bool_columns_exclusive_to_existing_table: vector of booleans referring to columns
exclusive to the existing table
$vec_bool_columns_exclusive_to_incoming_table: vector of booleans referring to columns
exclusive to the incoming table
\item Err: dbError
}
}
\description{
Classify the columns of the incoming base or data table to check for intersection/s with existing
data and base data tables in the database
}
\examples{
list_fnames_tables = fn_simulate_tables(
     n_entries=50,
     n_dates=3,
     n_sites=3,
     n_treatments=3,
     n_loci=10e3,
     save_data_tables=TRUE)$list_fnames_tables
df = utils::read.delim(list_fnames_tables$fname_phenotypes, header=TRUE)
database = DBI::dbConnect(drv=RSQLite::SQLite(), dbname="test.sqlite")
### Prepare the tables
list_df_data_and_base_tables = fn_prepare_data_table_and_extract_base_tables(df=df, 
     database=database, table_name="phenotypes", verbose=TRUE)
df = list_df_data_and_base_tables$df_possibly_modified
df_first_half = droplevels(df[, 1:floor(ncol(df)/2)])
### Import the first half of the data table into the database
DBI::dbWriteTable(conn=database, name="phenotypes", value=df_first_half)
### Detect column intersections
list_set_classification_of_columns = fn_set_classification_of_columns(df=df, database=database, 
     table_name="phenotypes", verbose=TRUE)
DBI::dbDisconnect(database)
unlink("test.sqlite")
}
