% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{fn_simulate_tables}
\alias{fn_simulate_tables}
\title{Simulate base and data tables and save the data tables as tab-delimited and multi-sheet Excel files}
\usage{
fn_simulate_tables(
  n_entries = 100,
  n_dates = 10,
  n_sites = 10,
  n_treatments = 10,
  n_traits = 50,
  n_replications = 3,
  n_sensors = 3,
  n_abiotics = 10,
  n_loci = 1e+05,
  n_chromosomes = 7,
  save_data_tables = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{n_entries}{number of entries or samples or genotypes (Default=100)}

\item{n_dates}{number of measurement dates (Default=10)}

\item{n_sites}{number of measurement sites (Default=10)}

\item{n_treatments}{number of treatments (Default=10)}

\item{n_traits}{number of traits (Default=50)}

\item{n_replications}{number of replications (Default=3)}

\item{n_sensors}{number of sensors for measuring the abiotic factors (Default=3)}

\item{n_abiotics}{number of abiotic factors (Default=10)}

\item{n_loci}{number of loci (Default=100e3)}

\item{n_chromosomes}{number of chromosomes (Default=7)}

\item{save_data_tables}{save the data tables as tab-delimited and Excel files? (Default=FALSE)}

\item{verbose}{show simulation messages? (Default=TRUE)}
}
\value{
\itemize{
\item Ok:
\itemize{
\item $vec_ids: vector of sample/entry/pool names (same order as in G)
\item $vec_chr: vector of chromosome or scaffold names (same order as in G)
\item $vec_pos: vector of numeric positions per chromosome (same order as in G)
\item $vec_all: vector of allele names (same order as in G)
\item $df_entries: entries base table
\item $df_dates: dates base table
\item $df_sites: sites base table
\item $df_treatments: treatments base table
\item $df_traits: traits base table
\item $df_abiotics: abiotics base table
\item $df_loci: loci base table
\item $df_phenotypes: phenotypes data table
\item $df_environments: environments data table
\item $df_genotypes: genotypes data table
\item $list_fnames_tables
\itemize{
\item $fname_phenotypes: filename of tab-delimited phenotypes data table
\item $fname_environments: filename of tab-delimited environments data table
\item $fname_genotypes: filename of tab-delimited genotypes data table
\item $fname_data_tables: filename of Excel file consisting of one sheet each of the phenotypes, environments, and genotype data tables
}
}
\item Err: dbError
}
}
\description{
Simulate base and data tables and save the data tables as tab-delimited and multi-sheet Excel files
}
\examples{
list_sim = fn_simulate_tables(save_data_tables=FALSE)
list_fnames_tables = fn_simulate_tables(save_data_tables=TRUE)$list_fnames_tables
}
