% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{fn_update_database}
\alias{fn_update_database}
\title{Update the database}
\usage{
fn_update_database(fname_db, df, table_name, verbose = TRUE)
}
\arguments{
\item{fname_db}{name of the SQLite database file}

\item{df}{data frame representing a base (entries, dates, sites, treatments, traits, abiotics,
and loci) or data (phenotypes, environments and genotypes) table
Note that if the table is a genotypes tables then we expect \code{df} to be a data frame corresponding
to an allele frequency table, where the first 3 columns refer to the CHROMOSOME,
POSITION_PER_CHROMOSOME, and ALLELE, followed by the allele frequencies per sample,
entry or pool with their corresponding entry names as the column names.}

\item{table_name}{name of the base or data table represented by df}

\item{verbose}{Show messages? (Default=TRUE)}
}
\value{
\itemize{
\item Ok: 0
\item Err: dbError
}
}
\description{
Update the database
}
\examples{
### Initialise the database
fname_db = "test.sqlite"
fname_data_tables = fn_simulate_tables(
     n_entries=50,
     n_dates=3,
     n_sites=3,
     n_treatments=3,
     n_loci=10e3,
     save_data_tables=TRUE)$list_fnames_tables$fname_data_tables
list_df_data_tables = list()
for (table_name in c("phenotypes", "environments", "genotypes")) {
    eval(parse(text=paste0("list_df_data_tables$df_", table_name, 
    " = as.data.frame(readxl::read_excel(path=fname_data_tables, sheet=table_name))")))
}
fn_initialise_db(fname_db=fname_db, list_df_data_tables=list_df_data_tables)
### Update the database with new data
list_sim = fn_simulate_tables(
     n_entries=50,
     n_dates=3,
     n_sites=3,
     n_treatments=3,
     n_loci=10e3,
     save_data_tables=TRUE)
fn_update_database(fname_db=fname_db, df=list_sim$df_phenotypes, 
     table_name="phenotypes", verbose=TRUE)
unlink("test.sqlite")
}
