% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.new.site.R
\name{generate_site_id}
\alias{generate_site_id}
\title{Generate a unique site ID from latitude/longitude}
\usage{
generate_site_id(lat, lon)
}
\arguments{
\item{lat}{Latitude value}

\item{lon}{Longitude value}
}
\value{
A unique string ID generated by hashing the lat/lon values
}
\description{
Generate a unique site ID from latitude/longitude
}
\details{
This function is intended for database-less runs only by generating
         a unique site ID by hashing the rounded lat/lon coordinates.
}
\author{
Abhinav Pandey
}
