% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_db.R
\name{get.id}
\alias{get.id}
\title{get.id}
\usage{
get.id(table, colnames, values, con, create = FALSE, dates = FALSE)
}
\arguments{
\item{table}{name of table}

\item{colnames}{names of one or more columns used in where clause}

\item{values}{values to be queried in fields corresponding to colnames}

\item{con}{database connection object,}

\item{create}{logical: make a record if none found?}

\item{dates}{logical: update created_at and updated_at timestamps? Used only if `create` is TRUE}
}
\value{
will numeric
}
\description{
Retrieve id from a table matching query
}
\examples{
\dontrun{
pftid <- get.id("pfts", "name", "salix", con)
pftid <- get.id("pfts", c("name", "modeltype_id"), c("ebifarm.salix", 1), con)
}
}
\author{
David LeBauer
}
