% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.machine.info.R
\name{get.machine.host}
\alias{get.machine.host}
\title{Helper Function to retrieve machine host and machine informations}
\usage{
get.machine.host(host, con = NULL)
}
\arguments{
\item{host}{host information}

\item{con}{database connection}
}
\value{
list of machine host and machine information
}
\description{
Helper Function to retrieve machine host and machine informations
}
\author{
Abhinav Pandey
}
