% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.machine.info.R
\name{get.machine.info}
\alias{get.machine.info}
\title{Get machine information from db}
\usage{
get.machine.info(host, input.args, input.id = NULL, con = NULL)
}
\arguments{
\item{host}{host information}

\item{input.args}{input args for existing records}

\item{input.id}{input id for existing records}

\item{con}{database connection}
}
\value{
list of machine, input, and dbfile records
}
\description{
Get machine information from db
}
\author{
Betsy Cowdery, Michael Dietze, Ankur Desai, Tony Gardella, Luke Dramko
}
