% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.trait.data.R
\name{get.trait.data}
\alias{get.trait.data}
\title{Get trait data from the database.}
\usage{
get.trait.data(
  pfts,
  modeltype,
  dbfiles,
  database,
  forceupdate,
  trait.names = NULL
)
}
\arguments{
\item{pfts}{the list of pfts to get traits for}

\item{modeltype}{type of model that is used, this is is used to distinguis between different pfts with the same name.}

\item{dbfiles}{location where previous results are found}

\item{database}{database connection parameters (see `params`
argument to [db.query()])}

\item{forceupdate}{(Logical) If `TRUE`, force a database update
whether or not it is needed. If `FALSE`, only update if an
update is needed.}

\item{trait.names}{Character vector of trait names to search. If
`NULL` (default), use all traits that have a prior for at least
one of the `pfts`.}
}
\value{
list of PFTs with update posteriorids
}
\description{
This will use the following items from settings:
- `settings$pfts`
- `settings$model$type`
- `settings$database$bety`
- `settings$database$dbfiles`
- `settings$meta.analysis$update`
}
\author{
David LeBauer, Shawn Serbin, Alexey Shiklomanov
}
