% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_machine_info.R
\name{get_machine_host}
\alias{get_machine_host}
\title{Helper Function to retrieve machine host and machine informations}
\usage{
get_machine_host(host, con)
}
\arguments{
\item{host}{host information}

\item{con}{database connection}
}
\value{
list of machine host and machine information
}
\description{
Helper Function to retrieve machine host and machine informations
}
\author{
Abhinav Pandey
}
