% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.dplyr.R
\name{get_run_ids}
\alias{get_run_ids}
\title{Get vector of run IDs for a given workflow ID}
\usage{
get_run_ids(bety, workflow_id)
}
\arguments{
\item{bety}{BETYdb connection, as opened by `betyConnect()`}

\item{workflow_id}{Workflow ID}
}
\description{
Get vector of run IDs for a given workflow ID
}
