% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.dplyr.R
\name{get_users}
\alias{get_users}
\title{Get data frame of users and IDs}
\usage{
get_users(bety, session)
}
\arguments{
\item{bety}{BETYdb connection, as opened by `betyConnect()`}

\item{session}{Session object passed through Shiny}
}
\description{
Get data frame of users and IDs
}
