% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.dplyr.R
\name{get_var_names}
\alias{get_var_names}
\title{Get vector of variable names for a particular workflow and run ID}
\usage{
get_var_names(bety, workflow_id, run_id, remove_pool = TRUE)
}
\arguments{
\item{bety}{BETYdb connection, as opened by `betyConnect()`}

\item{workflow_id}{Workflow ID}

\item{run_id}{Run ID}

\item{remove_pool}{logical: ignore variables with 'pools' in their names?}
}
\description{
Get vector of variable names for a particular workflow and run ID
}
