% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{GLOBAL_df_valid_tables}
\alias{GLOBAL_df_valid_tables}
\title{Global variable encapsulating the valid table names, their classifications, and source table for the base tables}
\usage{
GLOBAL_df_valid_tables()
}
\description{
Global variable encapsulating the valid table names, their classifications, and source table for the base tables
}
\details{
Note that the genotypes table usually comes from an allele frequency table file (for format details, see: https://github.com/jeffersonfparil/imputef?tab=readme-ov-file#allele-frequency-table-tsv)
}
