% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{GLOBAL_list_required_colnames_per_table}
\alias{GLOBAL_list_required_colnames_per_table}
\title{Global variable listing the required column names for each base and data tables}
\usage{
GLOBAL_list_required_colnames_per_table()
}
\description{
Global variable listing the required column names for each base and data tables
}
\details{
Note that we do not include the genotypes table here because the columns required is dynamic, i.e. depend on the genome coverage.
}
