% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_table.R
\name{insert_table}
\alias{insert_table}
\title{Insert R data frame into SQL database}
\usage{
insert_table(values, table, con)
}
\arguments{
\item{values}{`data.frame` of values to write to SQL database}

\item{table}{Name of target SQL table, as character}

\item{con}{database connection object}
}
\value{
data frame with query results
}
\description{
First, subset to matching columns. Then, build an SQL string for the insert 
statement. Finally, insert into the database.
}
\examples{
library(dplyr)
irisfile <- tempfile(fileext = ".sqlite")
irisdb <- src_sqlite(irisfile, create = TRUE)
copy_to(irisdb, iris[1,], name = "iris", overwrite = TRUE)
insert_table(iris[-1,], "iris", irisdb$con)
tbl(irisdb, "iris")
}
