% Generated by roxygen2 (4.0.1): do not edit by hand
\name{killdbcons}
\alias{killdbcons}
\title{Kill existing database connections}
\value{
nothing, as a side effect closes all open connections
}
\description{
Kill existing database connections
}
\details{
Deprecated, this should never be called

resolves (provides workaround to) bug #769 caused by having too many open connections \code{Error during wrapup: RS-DBI driver: (cannot allocate a new connection -- maximum of 16 connections already opened)}
}
\author{
David LeBauer
}

