% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.dplyr.R
\name{load_data_single_run}
\alias{load_data_single_run}
\title{Load data for a single run of the model}
\usage{
load_data_single_run(bety, workflow_id, run_id)
}
\arguments{
\item{bety}{BETYdb connection, as opened by `betyConnect()`}

\item{workflow_id}{Workflow ID}

\item{run_id}{Run ID}
}
\description{
Load data for a single run of the model
}
