% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_table.R
\name{match_colnames}
\alias{match_colnames}
\title{Match names of local data frame to SQL table}
\usage{
match_colnames(values, table, con)
}
\arguments{
\item{values}{`data.frame` of values to write to SQL database}

\item{table}{Name of target SQL table, as character}

\item{con}{Database connection object}
}
\description{
Match names of local data frame to SQL table
}
