% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_table.R
\name{match_dbcols}
\alias{match_dbcols}
\title{Match column names and classes between local and SQL table}
\usage{
match_dbcols(values, table, con, coerce_col_class = TRUE, drop = TRUE)
}
\arguments{
\item{values}{`data.frame` of values to write to SQL database}

\item{table}{Name of target SQL table, as character}

\item{con}{Database connection object}

\item{coerce_col_class}{logical, whether or not to coerce local data columns 
to SQL classes. Default = `TRUE.`}

\item{drop}{logical. If `TRUE` (default), drop columns not found in SQL table.}
}
\value{
`values` `data.frame` with column names and classes matched to SQL
}
\description{
Match column names and classes between local and SQL table
}
