% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_inputs.R
\name{met_inputs}
\alias{met_inputs}
\title{Retrieve available met inputs for the given site, model, and hostname}
\usage{
met_inputs(dbcon, site_id, model_id, hostname)
}
\arguments{
\item{dbcon}{Database connection object}

\item{site_id}{Site ID (from `sites` table)}

\item{model_id}{Model ID (from `models`) table}

\item{hostname}{Hostname of machine}
}
\value{
`data.frame` of available met inputs
}
\description{
This is identical to the query performed by `web/03-inputs.php` to
populate the list of available sites. It may be useful for
debugging, or otherwise replicating PEcAn web interface behavior.
}
\author{
Alexey Shiklomanov
}
