% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.dplyr.R
\name{ncdays2date}
\alias{ncdays2date}
\title{Convert netcdf number of days to date}
\usage{
ncdays2date(time, unit)
}
\arguments{
\item{time}{time if dimension requests it}

\item{unit}{CF style unit (e.g. "days since 2010-01-01")}
}
\description{
Convert netcdf number of days to date
}
