% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz.imports.R
\docType{import}
\name{otherimports}
\alias{otherimports}
\alias{\%>\%}
\title{Imports from other packages}
\description{
Imports from other packages
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}

