% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{PEcAn.DB}
\alias{PEcAn.DB}
\alias{PEcAn.DB-package}
\title{Database functions for PEcAn, the Predictive Ecosystem Analyzer}
\description{
This package provides an interface between PEcAn and the BETY database.
For usage examples, please see \code{vignette("betydb_access")}
}
