% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priordupe.R
\name{priordupe}
\alias{priordupe}
\title{Duplicate existing prior for new pft}
\usage{
priordupe(parent.pft.name, new.pft.name, new.pft.definition, settings)
}
\arguments{
\item{parent.pft.name}{name of PFT to duplicate}

\item{new.pft.name}{name for new PFT. Must not be the same as parent.pft.name}

\item{new.pft.definition}{text for the new PFT's definition field.}

\item{settings}{PEcAn settings list, used only for Bety connection parameters}
}
\value{
ID of the newly created pft in database
}
\description{
Creates a new pft that is a duplicate of an existing pft,
including relationships with priors and species of the existing pft
}
\examples{
\dontrun{
priordupe(parent.pft.name    = "tempdecid",
          new.pft.name       = "mytempdecid",
          new.pft.definition = "mytempdecid is a new pft")
}
}
\author{
David LeBauer, Chris Black
}
