% Generated by roxygen2 (4.0.1): do not edit by hand
\name{query.base}
\alias{query.base}
\title{Query database}
\usage{
query.base(query, con = NULL, ...)
}
\arguments{
\item{query}{SQL query string}

\item{con}{database connection object}

\item{...}{optional arguments for connecting to database (e.g. password, user name, database)}
}
\value{
data frame with query results
}
\description{
Generic function to query trait database
}
\details{
Given a connection and a query, will return a query as a data frame

Deprecated, please use db.query
}
\examples{
\dontrun{
query.base('select count(id) from traits;')
}
}

