% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/query.base.R
\name{query.close}
\alias{query.close}
\title{Close existing database connections}
\usage{
query.close(con)
}
\arguments{
\item{con}{database connection object}
}
\value{
nothing, as a side effect closes all open connections
}
\description{
Close database connection
}
\details{
Deprecated, please use db.close

Closes a database connection
}
\author{
Rob Kooper
}

