\name{query.covariates}
\alias{query.covariates}
\title{Queries covariates from database for a given vector of trait id's}
\usage{
  query.covariates(trait.ids, con = query.bety.con(...),
    ...)
}
\arguments{
  \item{trait.ids}{list of trait ids}
}
\description{
  Queries covariates from database for a given vector of
  trait id's
}

