% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.trait.data.R
\name{query.data}
\alias{query.data}
\title{Query data and transform stats to SE by calling \code{\link{fetch.stats2se}};}
\usage{
query.data(trait, spstr,
  extra.columns = "ST_X(ST_CENTROID(sites.geometry)) AS lon, ST_Y(ST_CENTROID(sites.geometry)) AS lat, ",
  con = NULL, store.unconverted = FALSE, ids_are_cultivars = FALSE,
  ...)
}
\arguments{
\item{trait}{trait to query from the database}

\item{spstr}{IDs of species to query from, as a single comma-separated string}

\item{extra.columns}{other query terms to pass in. If unspecified, retrieves latitude and longitude}

\item{con}{database connection}

\item{ids_are_cultivars}{if TRUE, ids is a vector of cultivar IDs, otherwise they are species IDs}

\item{...}{extra arguments}
}
\description{
Function to query data from database for specific species and convert stat to SE
}
\seealso{
used in \code{\link{query.trait.data}}; \code{\link{fetch.stats2se}}; \code{\link{transformstats}} performs transformation calculations
}
\author{
David LeBauer, Carl Davidson
}
