% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.file.path.R
\name{query.file.path}
\alias{query.file.path}
\title{Get file path given id and machine}
\usage{
query.file.path(input.id, host_name, con)
}
\arguments{
\item{input.id}{The database id of the input file of the parent of the file being processed here. The parent will have the same data, but in a different format.}

\item{host_name}{character: machine where the file lives}

\item{con}{: database connection}
}
\description{
Get file path given id and machine
}
\author{
Betsy Cowdery
}
