% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.format.vars.R
\name{query.format.vars}
\alias{query.format.vars}
\title{Given input_id, return formats table and table of variables and units}
\usage{
query.format.vars(bety, input.id = NA, format.id = NA, var.ids = NA)
}
\arguments{
\item{bety}{database connection}

\item{input.id}{The database id of the input file of the parent of the file being processed here.  The parent will have the same data, but in a different format.}

\item{format.id}{numeric. Defaults to format.id if both provided}

\item{var.ids}{optional vector of variable IDs. If provided, limits results to these variables}
}
\description{
Given input_id, return formats table and table of variables and units
}
\author{
Betsy Cowdery, Ankur Desai, Istem Fer
}
