% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.pft.R
\name{query.pft_cultivars}
\alias{query.pft_cultivars}
\title{Select cultivars associated with a PFT}
\usage{
query.pft_cultivars(pft, modeltype = NULL, con)
}
\arguments{
\item{pft}{string pft name}

\item{modeltype}{type of model that is used, this is used to distinguish between different pfts with the same name.}

\item{con}{database connection}
}
\value{
tibble containing names and ids for each cultivar
  and the species it comes from
}
\description{
Given a PFT name and optionally a modeltype, finds its pft_id and
returns the cultivars associated with it.
}
\details{
A PFT is allowed to have associated species or associated
cultivars, but not both. If this function returns no results, try checking
your PFT with \code{\link{query.pft_species}} instead.
Note that the cultivars associated with one PFT *are* allowed to come from
multiple species, if desired.
}
