% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.pft.R
\name{query.pft_species}
\alias{query.pft_species}
\title{Query species given pft name}
\usage{
query.pft_species(pft, modeltype = NULL, con)
}
\arguments{
\item{pft}{string pft name}

\item{modeltype}{type of model that is used, this is used to distinguish between different pfts with the same name.}

\item{con}{database connection}
}
\value{
data.frame containing id, genus, species, scientificname of each species associated with pft
}
\description{
select plant id's associated with pft
}
\examples{
\dontrun{
  con <- db.open(list(host = "postgres", user = "bety", password = "carya")) 
  query.pft_species("ebifarm.pavi", con = con)
}
}
\author{
David LeBauer
}
