% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.site.R
\name{query.site}
\alias{query.site}
\title{Given site_id, return site table}
\usage{
query.site(site_id, dbcon)
}
\arguments{
\item{site_id}{numeric}

\item{dbcon}{: database connection}
}
\description{
Given site_id, return site table
}
\author{
Betsy Cowdery
}
