% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.yields.R
\name{query.yields}
\alias{query.yields}
\title{Query yield data and transform stats to SE by calling \code{\link{fetch.stats2se}};}
\usage{
query.yields(
  trait = "yield",
  spstr,
  extra.columns = "",
  con = NULL,
  ids_are_cultivars = FALSE,
  ...
)
}
\arguments{
\item{trait}{yield trait to query}

\item{spstr}{species to query for yield data}

\item{extra.columns}{other query terms to pass in. Optional}

\item{con}{database connection}

\item{ids_are_cultivars}{if TRUE, spstr contains cultivar IDs, otherwise they are species IDs}

\item{...}{extra arguments}
}
\description{
Function to query yields data from database for specific species and convert stat to SE
}
\seealso{
used in \code{\link{query.trait.data}}; \code{\link{fetch.stats2se}}; \code{\link{transformstats}} performs transformation calculations
}
\author{
<unknown>
}
