% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.prior.R
\name{query_priors}
\alias{query_priors}
\title{Query priors using prepared statements}
\usage{
query_priors(pft_names = NULL, traits = NULL, pft_ids = NULL,
  expand = TRUE, ...)
}
\arguments{
\item{pfts}{Character vector of PFT names}
}
\description{
Query priors using prepared statements
}
