\name{rename.jags.columns}
\alias{rename.jags.columns}
\title{\code{rename.jags.columns} renames the variables within output data frame trait.data}
\usage{
rename.jags.columns(data)
}
\arguments{
  \item{data}{data frame to with variables to rename}
}
\description{

}
\seealso{
used with \code{\link{jagify}};
}

