% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.dplyr.R
\name{runs}
\alias{runs}
\title{Get table of runs corresponding to a workflow}
\usage{
runs(bety, workflow_id)
}
\arguments{
\item{bety}{BETYdb connection, as opened by `betyConnect()`}

\item{workflow_id}{Workflow ID}
}
\description{
Get table of runs corresponding to a workflow
}
