% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_references.R
\name{search_reference_single}
\alias{search_reference_single}
\title{Perform crossref search for a single reference}
\usage{
search_reference_single(query, limit = 1, min_score = 85)
}
\arguments{
\item{query}{Citation string (length 1) to search for DOI}

\item{limit}{Number of results to return}

\item{min_score}{Minimum match score. Default (85) is fairly strict.}
}
\value{
`data.frame` containing crossref information converted to match bety citations table.
}
\description{
Requires the `rcrossref` package.
}
