% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-sqlar.R
\name{sqlar_compress}
\alias{sqlar_compress}
\title{Compress data using gzip}
\usage{
sqlar_compress(raw)
}
\arguments{
\item{raw}{An uncompressed byte vector.}
}
\value{
The compressed byte vector.
}
\description{
Compress data using gzip
}
