% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-sqlar.R
\name{sqlar_uncompress}
\alias{sqlar_uncompress}
\title{Uncompress gzip'd data}
\usage{
sqlar_uncompress(raw, sz)
}
\arguments{
\item{raw}{A compressed byte vector.}

\item{sz}{The size of the uncompressed byte vector.}
}
\value{
The uncompressed byte vector.
}
\description{
Uncompress gzip'd data
}
