% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.dplyr.R
\name{var_names_all}
\alias{var_names_all}
\title{Get vector of variable names for a particular workflow and run ID}
\usage{
var_names_all(bety, workflow_id, run_id)
}
\arguments{
\item{bety}{BETYdb connection, as opened by `betyConnect()`}

\item{workflow_id}{Workflow ID}

\item{run_id}{Run ID}
}
\description{
Get vector of variable names for a particular workflow and run ID
}
