% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vec_words}
\alias{vec_words}
\title{List of words found in Unix and Unix-like operating systems
File location: /usr/share/dict/words or /usr/dict/words}
\format{
\subsection{\code{vec_words}}{

A vector with 451,664 string elements.
}
}
\usage{
vec_words
}
\description{
List of words found in Unix and Unix-like operating systems
File location: /usr/share/dict/words or /usr/dict/words
}
\keyword{datasets}
