% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-vtable.R
\docType{class}
\name{virtualtable-class}
\alias{virtualtable-class}
\alias{VirtualTable}
\title{An S4 class to implement new virtual table modules}
\description{
An S4 class to implement new virtual table modules
}
\section{Slots}{

\describe{
\item{\code{database}}{The database connection. An S4 object of class "database".}

\item{\code{name}}{A character string. What to call the virtual table module.}

\item{\code{method}}{A function to register user-defined virtual table methods.}

\item{\code{.methodEnv}}{An environment in which virtual table methods are invoked.}
}}

\seealso{
\code{\link{new.virtualtable}}
}
