% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.dplyr.R
\name{workflow}
\alias{workflow}
\title{Get single workflow by workflow_id}
\usage{
workflow(bety, workflow_id)
}
\arguments{
\item{bety}{BETYdb connection, as opened by `betyConnect()`}

\item{workflow_id}{Workflow ID}
}
\description{
Get single workflow by workflow_id
}
