% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.dplyr.R
\name{workflows}
\alias{workflows}
\title{list of workflows that exist}
\usage{
workflows(bety, ensemble = FALSE)
}
\arguments{
\item{bety}{BETYdb connection, as opened by `betyConnect()`}

\item{ensemble}{Logical. Use workflows from ensembles table.}
}
\description{
list of workflows that exist
}
