% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/query.trait.data.R
\name{arrhenius.scaling.traits}
\alias{arrhenius.scaling.traits}
\title{Function to apply Arrhenius scaling to 25 degC for temperature-dependent traits}
\usage{
arrhenius.scaling.traits(data, covariates, temp.covariates, new.temp = 25,
  missing.temp = 25)
}
\arguments{
\item{data}{data frame of data to scale, as returned by query.data()}

\item{covariates}{data frame of covariates, as returned by query.covariates().
Note that data with no matching covariates will be unchanged.}

\item{temp.covariates}{names of covariates used to adjust for temperature;
if length > 1, order matters (first will be used preferentially)}

\item{new.temp}{the reference temperature for the scaled traits. Curerntly 25 degC}

\item{missing.temp}{the temperature assumed for traits with no covariate found. Curerntly 25 degC}
}
\description{

}
\author{
Carl Davidson, David LeBauer, Ryan Kelly
}

