% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{db.close}
\alias{db.close}
\title{Close database connection}
\usage{
db.close(con)
}
\arguments{
\item{con}{database connection to be closed}
}
\value{
connection to database
}
\description{
Generic function to close a database connection
}
\details{
Close a previously opened connection to a database.
}
\examples{
\dontrun{
db.close(con)
}
}
\author{
Rob Kooper
}

