% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{db.exists}
\alias{db.exists}
\title{db.exists}
\usage{
db.exists(params, write = TRUE, table = NA)
}
\arguments{
\item{params}{database connection information}
}
\value{
TRUE if database connection works; else FALSE
}
\description{
Test connection to database
}
\details{
Useful to only run tests that depend on database when a connection exists
}
\author{
David LeBauer, Rob Kooper
}

