% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-sqlar.R
\name{db.unsqlar}
\alias{db.unsqlar}
\title{Unpack a SQLite archive.}
\usage{
db.unsqlar(db, path)
}
\arguments{
\item{db}{The database connection. S4 object of class "database".}

\item{path}{The path to unpack the archive under.}
}
\value{
None.
}
\description{
Unpack a SQLite archive.
}
\details{
This will read the sqlar table in the database and write out
its content into a filesystem hierarchy under \code{path}.
}
\note{
The \code{\link{sqlar_uncompress}} function must be registered
with the database connection for this function to work. This is done
automatically with \code{\link{db.open}}
}
