% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-httpd.R
\name{db.urlpath}
\alias{db.urlpath}
\title{Retrieve elements of a URL path}
\usage{
db.urlpath(path, what)
}
\arguments{
\item{path}{A URL path.}

\item{what}{The part of the path to retrieve ('name', 'head', 'tail', or 'info').}
}
\value{
A character string.
}
\description{
URL paths will all have the format /custom/<name>/<head>/<tail>
Note that tail may be multipart. That is, in the URL
/custom/foo/abc/def/hij the <name> = foo, <head> = abc and
<tail> = def/hij.
}
