% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfiles.R
\name{dbfile.check}
\alias{dbfile.check}
\title{Check for a file in the dbfiles tables}
\usage{
dbfile.check(type, container.id, con, hostname = PEcAn.remote::fqdn(),
  machine.check = TRUE, return.all = FALSE)
}
\arguments{
\item{type}{the type of dbfile (Input, Posterior)}

\item{container.id}{the id of container type}

\item{con}{database connection object}

\item{hostname}{the name of the host where the file is stored, this will default to the name of the current machine}

\item{machine.check}{setting to check for file on named host, otherwise will check for any file given container id}

\item{return.all}{logical flag, if TRUE will return all the files in the directory, not only the max(dbfiles$updated_at) subset}
}
\value{
data.frame with the id, filename and pathname of all the files that are associated
}
\description{
Function to check to see if a file exists in the dbfiles table
}
\details{
This will check the dbfiles and machines to see if the file exists
}
\examples{
\dontrun{
  dbfile.check('Input', 7, dbcon)
}
}
\author{
Rob Kooper
}
