% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfiles.R
\name{dbfile.check}
\alias{dbfile.check}
\title{List files associated with a container and machine exist in
`dbfiles` table}
\usage{
dbfile.check(
  type,
  container.id,
  con,
  hostname = PEcAn.remote::fqdn(),
  machine.check = TRUE,
  return.all = FALSE
)
}
\arguments{
\item{type}{The type of `dbfile`, as a character. Must be either
"Input", "Posterior", or "Model".}

\item{container.id}{the ID of container type. E.g. if `type` is
"Input", this will correspond to the `id` column from the
`inputs` table.}

\item{con}{database connection object}

\item{hostname}{the name of the host where the file is stored,
this will default to the name of the current machine}

\item{machine.check}{setting to check for file on named host,
otherwise will check for any file given container id}

\item{return.all}{(Logical) If `TRUE`, return all files. If
`FALSE`, return only the most recent files.}
}
\value{
`data.frame` with the id, filename and pathname of all the files that are associated
}
\description{
List files associated with a container and machine exist in
`dbfiles` table
}
\examples{
\dontrun{
  dbfile.check("Input", 7, dbcon)
}
}
\author{
Rob Kooper, Alexey Shiklomanov
}
