% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/query.trait.data.R
\name{filter.sunleaf.traits}
\alias{filter.sunleaf.traits}
\title{Function to filter out upper canopy leaves}
\usage{
filter.sunleaf.traits(data, covariates)
}
\arguments{
\item{data}{input data}

\item{covariates}{covariate data}
}
\description{

}
\author{
<unknown>
}

