\name{filter.sunleaf.traits}
\alias{filter.sunleaf.traits}
\title{Function to filter out upper canopy leaves}
\usage{
  filter.sunleaf.traits(data, covariates)
}
\arguments{
  \item{data}{input data}

  \item{covariates}{covariate data}
}
\description{
  Function to filter out upper canopy leaves
}
\author{
  <unknown>
}

