% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{fn_check_import_inputs}
\alias{fn_check_import_inputs}
\title{Check the validity of importation function inputs}
\usage{
fn_check_import_inputs(df, database, table_name, check_UIDs = FALSE)
}
\arguments{
\item{df}{number of entries or samples or genotypes (Default=100)}

\item{database}{number of measurement dates (Default=10)}

\item{table_name}{number of measurement sites (Default=10)}

\item{check_UIDs}{number of treatments (Default=10)}
}
\value{
\itemize{
\item Ok:
NULL
\item Err: dbError
}
}
\description{
Check the validity of importation function inputs
}
\examples{
list_fnames_tables = fn_simulate_tables(
     n_entries=50,
     n_dates=3,
     n_sites=3,
     n_treatments=3,
     n_loci=10e3,
     save_data_tables=TRUE)$list_fnames_tables
df = utils::read.delim(list_fnames_tables$fname_phenotypes, header=TRUE)
database = DBI::dbConnect(drv=RSQLite::SQLite(), dbname="test.sqlite")
null_error = fn_check_import_inputs(df=df, database=database, table_name="phenotypes", check_UIDs=FALSE)
non_null_error = fn_check_import_inputs(df=df[, -1:-5], database=database, table_name="phenotypes", check_UIDs=FALSE)
}
