% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{fn_remove_quotes_and_newline_characters_in_data}
\alias{fn_remove_quotes_and_newline_characters_in_data}
\title{Remove quotes and newline characters ("\n" in Linux and "\r\n" in Windows)}
\usage{
fn_remove_quotes_and_newline_characters_in_data(df, verbose = TRUE)
}
\arguments{
\item{df}{number of entries or samples or genotypes (Default=100)}

\item{db}{number of measurement dates (Default=10)}

\item{table_name}{number of measurement sites (Default=10)}

\item{check_UIDs}{number of treatments (Default=10)}
}
\value{
\itemize{
\item Ok:
NULL
\item Err: dbError
}
}
\description{
Remove quotes and newline characters ("\n" in Linux and "\r\n" in Windows)
}
\examples{
list_fnames_tables = fn_simulate_tables(
     n_entries=50,
     n_dates=3,
     n_sites=3,
     n_treatments=3,
     n_loci=10e3,
     save_data_tables=TRUE)$list_fnames_tables
df = utils::read.delim(list_fnames_tables$fname_phenotypes, header=TRUE)
db = DBI::dbConnect(drv=RSQLite::SQLite(), dbname="test.sqlite")
null_error = fn_check_import_inputs(df=df, db=db, table_name="phenotypes", check_UIDs=FALSE)
non_null_error = fn_check_import_inputs(df=df[, -1:-5], db=db, table_name="phenotypes", check_UIDs=FALSE)
}
