% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.id}
\alias{get.id}
\title{get.id}
\usage{
get.id(table, colnames, values, con)
}
\arguments{
\item{table}{name of table}

\item{colnames}{names of one or more columns used in where clause}

\item{values}{values to be queried in fields corresponding to colnames}
}
\value{
will numeric
}
\description{
Retrieve id from a table matching query
}
\examples{
pftid <- get.id("pfts", "name", "salix", con)
pftid <- get.id("pfts", c("name", "modeltype_id"), c("ebifarm.salix", 1), con)
}
\author{
David LeBauer
}

