% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{get.id}
\alias{get.id}
\title{get.id}
\usage{
get.id(table, colnames, values, con, create = FALSE, dates = FALSE)
}
\arguments{
\item{table}{name of table}

\item{colnames}{names of one or more columns used in where clause}

\item{values}{values to be queried in fields corresponding to colnames}
}
\value{
will numeric
}
\description{
Retrieve id from a table matching query
}
\examples{
\dontrun{
pftid <- get.id("pfts", "name", "salix", con)
pftid <- get.id("pfts", c("name", "modeltype_id"), c("ebifarm.salix", 1), con)
}
}
\author{
David LeBauer
}

