% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.new.site.R
\name{get.new.site}
\alias{get.new.site}
\title{Get New Site Info}
\usage{
get.new.site(site, con, latlon = NULL)
}
\arguments{
\item{site}{a dataframe with site information including id, lat, lon, and time_zone.}

\item{con}{Database connection object.}

\item{latlon}{Optional global latlon object which will store updated lat/lon.}
}
\value{
a dataframe with new site information on lat, lon and time_zone
}
\description{
Get new site info using provided site information
}
\examples{
get.new.site(site=data.frame(id=1,lat=40.1,lon=-88.2,time_zone="UTC"),con,latlon=NULL)
}
\author{
Abhinav Pandey
}
