% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.trait.data.pft.R
\name{get.trait.data.pft}
\alias{get.trait.data.pft}
\title{Get trait data from the database for a single PFT}
\usage{
get.trait.data.pft(
  pft,
  modeltype,
  dbfiles,
  dbcon,
  trait.names,
  forceupdate = FALSE,
  write = FALSE
)
}
\arguments{
\item{pft}{list of settings for the pft whose traits to retrieve. See details.}

\item{modeltype}{type of model that is used, this is used to distinguish
between different pfts with the same name.}

\item{dbfiles}{location where previous results are found}

\item{dbcon}{database connection}

\item{trait.names}{list of trait names to retrieve}

\item{forceupdate}{set this to true to force an update, auto will check to
see if an update is needed.}

\item{write}{(Logical) If \code{TRUE} updated posteriors will be written to
BETYdb.  Defaults to \code{FALSE}.}
}
\value{
The \code{pft} input list, updated with \code{pft$posteriorid} set to the
ID of the (possibly new) posterior record in BETYdb. Also contains
\code{pft$trait_data} and \code{pft$prior_distns} for in-memory chaining on both
the cache-hit and cache-miss paths. The posterior ID can be used to
locate the output files (\code{trait.data.Rdata}, \code{prior.distns.Rdata},
etc.) via BETYdb's \code{dbfiles} table.
}
\description{
Get trait data from the database for a single PFT
}
\details{
\code{pft} should be a list containing at least \code{name} and \code{outdir}, and
optionally \code{posteriorid} and \code{constants}.

Internally this wrapper delegates all database queries to
\code{\link{get_trait_data_pft}} exactly once.  The returned objects are
used for both the cache-staleness check and the save step, so the database
is never queried more than once per call regardless of whether the cache
hits or misses.
}
\author{
David LeBauer, Shawn Serbin, Rob Kooper
}
