% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.trait.data.pft.R
\name{get.trait.data.pft}
\alias{get.trait.data.pft}
\title{Get trait data from the database for a single PFT}
\usage{
get.trait.data.pft(pft, modeltype, dbfiles, dbcon, trait.names,
  forceupdate = FALSE)
}
\arguments{
\item{pft}{list of settings for the pft whose traits to retrieve. See details}

\item{modeltype}{type of model that is used, this is used to distinguish between different pfts with the same name.}

\item{dbfiles}{location where previous results are found}

\item{dbcon}{database connection}

\item{trait.names}{list of trait names to retrieve}

\item{forceupdate}{set this to true to force an update, auto will check to see if an update is needed.}
}
\value{
updated pft with posteriorid
}
\description{
Get trait data from the database for a single PFT
}
\details{
`pft` should be a list containing at least `name` and `outdir`, and optionally `posteriorid` and `constants`. BEWARE: All existing files in `outir` will be deleted!
}
\author{
David LeBauer, Shawn Serbin, Rob Kooper
}
