% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.dplyr.R
\name{get_workflow_ids}
\alias{get_workflow_ids}
\title{Get vector of workflow IDs}
\usage{
get_workflow_ids(bety, query, all.ids = FALSE)
}
\arguments{
\item{bety}{BETYdb connection, as opened by `betyConnect()`}

\item{query}{Named vector or list of workflow IDs}

\item{all.ids}{logical: return a list of all workflow_ids in the BETY database,
or just those that are part of the query?}
}
\description{
Get vector of workflow IDs
}
