% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.dplyr.R
\name{get_workflow_ids}
\alias{get_workflow_ids}
\title{Get vector of workflow IDs}
\usage{
get_workflow_ids(bety, query, all.ids = FALSE)
}
\arguments{
\item{bety}{BETYdb connection, as opened by `betyConnect()`}

\item{query}{Named vector or list of workflow IDs}
}
\description{
Get vector of workflow IDs
}
