% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.format.vars.R
\name{insert.format.vars}
\alias{insert.format.vars}
\title{Insert Format and Format-Variable Records}
\usage{
insert.format.vars(con, format_notes, format_name, header, skip, mimetype_id,
  formats_variables_df = NULL)
}
\arguments{
\item{con}{Bety connection object}

\item{format_notes}{Character string used to}

\item{format_name}{}

\item{header}{}

\item{skip}{}

\item{mimetype_id}{}

\item{formats_variables_df}{A 'data.frame' consisting of entries that correspond to columns in the formats-variables table}
}
\value{
format id
}
\description{
Insert Format and Format-Variable Records
}
\examples{
\dontrun{
bety <- betyConnect
  insert.format.vars(con = bety$con, formats_df, formats_variables_df)
}
}
\author{
Liam Burke (liam.burke24@gmail.com)
}
