% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.format.vars.R
\name{insert.format.vars}
\alias{insert.format.vars}
\title{Insert Format and Format-Variable Records}
\usage{
insert.format.vars(con, format_name, mimetype_id, notes = NULL,
  header = TRUE, skip = 0, formats_variables = NULL)
}
\arguments{
\item{con}{SQL connection to BETYdb}

\item{format_name}{The name of the format. Type: character string.}

\item{mimetype_id}{The id associated with the mimetype of the format. Type: integer.}

\item{notes}{Additional description of the format: character string.}

\item{header}{Boolean that indicates the presence of a header in the format. Defaults to "TRUE".}

\item{skip}{Integer that indicates the number of lines to skip in the header. Defaults to 0.}

\item{formats_variables}{A 'tibble' consisting of entries that correspond to columns in the formats-variables table. See Details for further information.}
}
\value{
format_id
}
\description{
Insert Format and Format-Variable Records
}
\examples{
\dontrun{
bety <- betyConnect()

formats_variables_tibble <- tibble::tibble(
       variable_id = c(411, 135, 382), # integer 
       name = c("NPP", NA, "YEAR"), # 
       unit = c("g C m-2 yr-1", NA, NA),
       storage_type = c(NA, NA, "\%Y"),
       column_number = c(NA, NA, NA),
 )
  insert.format.vars(con = bety$con, name = "LTER-HFR-103", mimetype_id = 1090, notes = "NPP from Harvard Forest.", header = FALSE, skip = 0, formats_variables = formats_variables_tibble)
}
}
\author{
Liam Burke (liam.burke24@gmail.com)
}
