% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.format.vars.R
\name{insert.format.vars}
\alias{insert.format.vars}
\title{Insert Format and Format-Variable Records}
\usage{
insert.format.vars(con, name, mimetype_id, notes = NULL, header = TRUE,
  skip = 0, formats_variables_df = NULL)
}
\arguments{
\item{con}{SQL connection to BETYdb}

\item{name}{The name of the format. Type: character string.}

\item{mimetype_id}{The id associated with the mimetype of the format. Type: integer.}

\item{notes}{Additional description of the format: character string.}

\item{header}{Boolean that indicates the presence of a header in the format. Defaults to "TRUE".}

\item{skip}{Integer that indicates the number of lines to skip in the header. Defaults to 0.}

\item{formats_variables_df}{A 'data.frame' consisting of entries that correspond to columns in the formats-variables table. See Details for further information.}
}
\value{
format_id
}
\description{
Insert Format and Format-Variable Records
}
\details{
It is very important that the formats_variables_df 'data.frame' be structured in a specific format so that the SQL query functions properly. First, all arguments should be passed as vectors so that each entry will correspond with a specific row.
\describe{
\item{variable_id}{(Required) Vector of integers.}
\item{name}{(Optional) Vector of character strings. Need only be specified if the} 
\item{unit}{(Optional) Vector of character strings. }
\item{storage_type}{(Optional) Vector of character strings}
\item{column_number}{(Optional) Vector of integers that determines the} 
}
}
\examples{
\dontrun{
bety <- betyConnect

formats_variables_df <- tibble::tibble(
       variable_id = c(327, 2.98e+02), # integer 
       name = c("New Name", "Other New Name"), # 
       unit = c("Yoonits", "Other Units"),
       storage_type = c("\%j", "\%Y"),
       column_number = c(as.integer(22), as.integer(24)),
 )
  insert.format.vars(con = bety$con, format_notes = "Info about format", format_name = "New Format", header = TRUE, skip = 2, mimetype_id = 1090, formats_variables_df)
}
}
\author{
Liam Burke (liam.burke24@gmail.com)
}
