% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.format.vars.R
\name{insert.format.vars}
\alias{insert.format.vars}
\title{Insert Format and Format-Variable Records}
\usage{
insert.format.vars(con, formats_df, formats_variables_df = NULL)
}
\arguments{
\item{con}{Bety connection object}

\item{formats_df}{A 'data.frame' consisting of entries that correspond to columns in the formats table}

\item{formats_variables_df}{A 'data.frame' consisting of entries that correspond to columns in the formats-variables table}
}
\value{
Data frame: Inner join of SQL table and input data frame (as unevaluated "lazy query" table)
}
\description{
Insert Format and Format-Variable Records
}
\examples{
\dontrun{
bety <- betyConnect
  insert.format.vars(con = bety$con, formats_df, formats_variables_df)
}
}
\author{
Liam Burke (liam.burke24@gmail.com)
}
