% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_table.R
\name{insert_table}
\alias{insert_table}
\title{Insert R data frame into SQL database}
\usage{
insert_table(values, table, con, coerce_col_class = TRUE, drop = TRUE)
}
\arguments{
\item{values}{`data.frame` of values to write to SQL database}

\item{table}{Name of target SQL table, as character}

\item{con}{Database connection object}

\item{coerce_col_class}{logical, whether or not to coerce local data columns 
to SQL classes. Default = `TRUE.`}

\item{drop}{logical. If `TRUE` (default), drop columns not found in SQL table.}
}
\value{
data frame with query results
}
\description{
First, subset to matching columns. Then, make sure the local and SQL column 
classes match, coercing local to SQL as necessary (or throwing an error). 
Then, build an SQL string for the insert statement. Finally, insert into the 
database.
}
\examples{
irisdb <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
dplyr::copy_to(irisdb, iris[1,], name = "iris", overwrite = TRUE)
insert_table(iris[-1,], "iris", irisdb)
dplyr::tbl(irisdb, "iris")
}
