% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.dplyr.R
\name{ncdays2date}
\alias{ncdays2date}
\title{Convert netcdf number of days to a datetime}
\usage{
ncdays2date(time, unit)
}
\arguments{
\item{time}{number of time units elapsed since origin}

\item{unit}{string containing CF-style time unit including origin (e.g. "days since 2010-01-01")}
}
\description{
Convert netcdf number of days to a datetime
}
