% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEcAn.DB-package.R
\docType{package}
\name{PEcAn.DB-package}
\alias{PEcAn.DB}
\alias{PEcAn.DB-package}
\title{Database functions for PEcAn, the Predictive Ecosystem Analyzer}
\description{
This package provides an interface between PEcAn and the BETY database.
For usage examples, please see \code{vignette("betydb_access")}
}
\author{
\strong{Maintainer}: David LeBauer \email{dlebauer@email.arizona.edu}

Authors:
\itemize{
  \item Mike Dietze \email{dietze@bu.edu}
  \item Rob Kooper \email{kooper@illinois.edu}
  \item Shawn Serbin \email{sserbin@bnl.gov}
  \item Elizabeth Cowdery \email{ecowdery@bu.edu}
  \item Ankur Desai \email{desai@aos.wisc.edu}
  \item Istem Fer \email{istem.fer@fmi.fi}
  \item Alexey Shiklomanov \email{alexey.shiklomanov@pnnl.gov}
  \item Tony Gardella \email{tonygard@bu.edu}
  \item Chris Black \email{chris@ckblack.org}
  \item Liam Burke \email{liam.burke24@gmail.com}
  \item Ryan Kelly \email{rykelly@bu.edu}
  \item Dan Wang
  \item Carl Davidson \email{davids14@illinois.edu}
  \item Xiaohui Feng \email{feng22@illinois.edu}
  \item Shashank Singh \email{shashanksingh819@gmail.com}
}

Other contributors:
\itemize{
  \item University of Illinois, NCSA [copyright holder]
}

}
\keyword{internal}
