% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priordupe.R
\name{priordupe}
\alias{priordupe}
\title{Duplicate PFT}
\usage{
priordupe(parent.pft.name = NULL, new.pft.name = NULL,
  new.pft.definition = NULL, settings = NULL)
}
\arguments{
\item{parent.pft.name}{name of PFT to duplicate}

\item{new.pft.name}{name for new PFT. Must not be the same as parent.pft.name}

\item{new.pft.definition}{text for the new PFT's definition field. If NULL the definition will be empty, not copied from the parent PFT.}
}
\value{
nothing, creates new pft in database as a side-effect
}
\description{
Duplicate existing prior for new pft
}
\details{
Creates a new pft that is a duplicate of an existing pft,
including relationships with priors and species of the existing pft
}
\examples{
\dontrun{
priordupe(parent.pft.name    = "tempdecid",
          new.pft.name       = "mytempdecid",
          new.pft.definition = "mytempdecid is a new pft")
}
}
\author{
David LeBauer
}
