\name{query.covariates}
\alias{query.covariates}
\title{Queries covariates from database for a given vector of trait id's}
\usage{
  query.covariates(trait.ids,
    con = query.base.con(settings), ...)
}
\arguments{
  \item{trait.ids}{list of trait ids}

  \item{con}{database connection}

  \item{...}{extra arguments}
}
\description{
  Queries covariates from database for a given vector of
  trait id's
}
\author{
  <unknown>
}

