% Generated by roxygen2 (4.0.1): do not edit by hand
\name{query.covariates}
\alias{query.covariates}
\title{Queries covariates from database for a given vector of trait id's}
\arguments{
\item{trait.ids}{list of trait ids}

\item{con}{database connection}

\item{...}{extra arguments}
}
\description{

}
\author{
<unknown>
}

