% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.trait.data.R
\name{query.covariates}
\alias{query.covariates}
\title{Queries covariates from database for a given vector of trait id's}
\usage{
query.covariates(trait.ids, con = NULL, ...)
}
\arguments{
\item{trait.ids}{list of trait ids}

\item{con}{database connection}

\item{...}{extra arguments}
}
\author{
<unknown>
}

