% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.file.path.R
\name{query.file.path}
\alias{query.file.path}
\title{Get file path given id and machine}
\usage{
query.file.path(input.id, host_name, con)
}
\arguments{
\item{host_name}{character: machine where the file lives}

\item{con}{: database connection}

\item{input_id}{numeric: BETY id of the file ("container") to find}
}
\author{
Betsy Cowdery
}
