% Generated by roxygen2 (4.0.2): do not edit by hand
\name{query.format.vars}
\alias{query.format.vars}
\title{Given input_id, return formats table and table of variables and units}
\usage{
query.format.vars(input.id, con)
}
\arguments{
\item{con}{: database connection}

\item{input_id}{}
}
\description{
Given input_id, return formats table and table of variables and units
}
\author{
Betsy Cowdery
}

