% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.format.vars.R
\name{query.format.vars}
\alias{query.format.vars}
\title{Given input_id, return formats table and table of variables and units}
\usage{
query.format.vars(bety, input.id = NA, format.id = NA, var.ids = NA)
}
\arguments{
\item{con}{: database connection}
}
\author{
Betsy Cowdery, Ankur Desai, Istem Fer
}
