% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.format.vars.R
\name{query.format.vars}
\alias{query.format.vars}
\title{Look up names and units of input variables from a format id or input id}
\usage{
query.format.vars(bety, input.id = NA, format.id = NA, var.ids = NA)
}
\arguments{
\item{bety}{database connection}

\item{input.id, format.id}{numeric. Defaults to format.id if both provided}

\item{var.ids}{optional vector of variable IDs. If provided, limits results to these variables}
}
\description{
Look up names and units of input variables from a format id or input id
}
\author{
Betsy Cowdery, Ankur Desai, Istem Fer
}
