% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.format.vars.R
\name{query.format.vars}
\alias{query.format.vars}
\title{Given input_id, return formats table and table of variables and units}
\usage{
query.format.vars(input.id, con)
}
\arguments{
\item{con}{: database connection}
}
\author{
Betsy Cowdery
}

