% Generated by roxygen2 (4.0.2): do not edit by hand
\name{query.pft_species}
\alias{query.pft_species}
\title{Query species given pft name}
\usage{
query.pft_species(pft, modeltype, con)
}
\arguments{
\item{pft}{string pft name}

\item{modeltype}{type of model that is used, this is is used to distinguis between different pfts with the same name.}

\item{con}{database connection}
}
\value{
string of species.id for species associated with pft
}
\description{
select plant id's associated with pft
}
\examples{
\dontrun{
query.pft_species('ebifarm.pavi')
query.pft_species(settings = read.settings("pecan.xml"))
}
}
\author{
David LeBauer
}

