% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.prior.R
\name{query.priors}
\alias{query.priors}
\title{Query Priors}
\usage{
query.priors(pft, trstr = NULL, con = NULL, ...)
}
\arguments{
\item{pft}{ID number of the PFT in the database}

\item{trstr}{String of traits to query priors for. If passed as a
character vector, it will be concatenated to a single string
using [PEcAn.utils::vecpaste()].}

\item{con}{Database connection object.}

\item{...}{Optional arguments for connecting to database (e.g.
password, user name, database).}
}
\value{
`data.frame` of priors for each trait and the given PFT.
}
\description{
Query priors associated with a plant functional type and a set of traits.
}
\details{
If neither `con` nor `...` are provided, this will try to
  connect to BETY using a `settings` object in the current
  environment.
}
\examples{
\dontrun{
  con <- db.open(...)
  query.priors("ebifarm.pavi", c("SLA", "Vcmax", "leaf_width"), con = con)
}
}
\author{
David LeBauer, Alexey Shiklomanov
}
