% Generated by roxygen2 (4.0.2): do not edit by hand
\name{query.site}
\alias{query.site}
\title{Given site_id, return site table}
\usage{
query.site(site.id, con)
}
\arguments{
\item{con}{: database connection}

\item{site_id}{}
}
\description{
Given site_id, return site table
}
\author{
Betsy Cowdery
}

