% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.site.R
\name{query.site}
\alias{query.site}
\title{Given site_id, return site table}
\usage{
query.site(site.id, con)
}
\arguments{
\item{con}{: database connection}

\item{site_id}{numeric}
}
\description{
Given site_id, return site table
}
\author{
Betsy Cowdery
}
