% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.traits.R
\name{query.traits}
\alias{query.traits}
\title{Query trait data}
\usage{
query.traits(ids, priors, con = NULL, update.check.only = FALSE,
  ids_are_cultivars = FALSE)
}
\arguments{
\item{ids}{vector of species or cultivar id's from trait database}

\item{priors}{vector of parameters for which priors have been specified}

\item{con}{database connection object}

\item{update.check.only}{if TRUE, returns results but does not print summaries}

\item{ids_are_cultivars}{if TRUE, ids is a vector of cultivar IDs, otherwise they are species IDs}
}
\value{
list of dataframes, each with data for one trait
}
\description{
Query available trait data associated with a given pft and a list of traits
}
\examples{
\dontrun{
species <- query.pft_species('ebifarm.c4crop')
spstr <- vecpaste(species$id)
trvec <- c('leafN', 'SLA')
trait.data <- query.traits(spstr, trvec)
}
}
\seealso{
\code{\link{query.trait.data}}
}
\author{
David LeBauer, Carl Davidson, Shawn Serbin
}
