% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_pfts.R
\name{query_pfts}
\alias{query_pfts}
\title{Retrieve PFT ID, name, and type from BETY}
\usage{
query_pfts(dbcon, pft_names, modeltype = NULL, strict = FALSE)
}
\arguments{
\item{dbcon}{Database connection object}

\item{pft_names}{character vector of PFT names}

\item{modeltype}{character.
If specified, only returns PFTs matching this modeltype.
If NULL, considers all modeltypes.}
}
\value{
`data.frame` containing PFT ID (`id`), type (`pft_type`),
  and name (`name`).
}
\description{
Retrieve PFT ID, name, and type from BETY
}
\author{
Alexey Shiklomanov, Chris Black
}
