% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric_setdiff.R
\name{symmetric_setdiff}
\alias{symmetric_setdiff}
\title{Symmetric set difference of two data frames}
\usage{
symmetric_setdiff(
  x,
  y,
  xname = "x",
  yname = "y",
  namecol = "source",
  simplify_types = TRUE
)
}
\arguments{
\item{x, y}{`data.frame`s to compare}

\item{xname}{Label for data in x but not y. Default = "x"}

\item{yname}{Label for data in y but not x. Default = "y"}

\item{namecol}{Name of label column. Default = "source".}

\item{simplify_types}{(Logical) If `TRUE`, coerce anything that
isn't numeric to character, to facilitate comparison.}
}
\value{
`data.frame` of data not common to x and y, with additional
  column (`namecol`) indicating whether data are only in x
  (`xname`) or y (`yname`)
}
\description{
Symmetric set difference of two data frames
}
\examples{
xdf <- data.frame(a = c("a", "b", "c"),
                  b = c(1, 2, 3),
                  stringsAsFactors = FALSE)
ydf <- data.frame(a = c("a", "b", "d"),
                  b = c(1, 2.5, 3),
                  stringsAsFactors = FALSE)
symmetric_setdiff(xdf, ydf)
}
