% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try2sqlite.R
\name{try2sqlite}
\alias{try2sqlite}
\title{Convert TRY text file to SQLite database}
\usage{
try2sqlite(try_files, sqlite_file = "try.sqlite")
}
\arguments{
\item{try_files}{Character vector of file names containing TRY data.
Multiple files are combined with `data.table::rbindlist`.}

\item{sqlite_file}{Target SQLite database file name, as character.}
}
\description{
The TRY file is huge and unnecessarily long, which makes it difficult to 
work with. The resulting SQLite database is much smaller on disk, and can be 
read much faster thanks to lazy evaluation.
}
\details{
The resulting TRY SQLite database contains the following tables:
 - `values` -- The actual TRY data. Links to all other tables through ID columns.
 - `traits` -- Description of trait and data names. Links to `values` through `DataID`. Similar to BETY `variables` table.
 - `datasets` -- Description of datasets and references/citations. Links to `values` through `DatasetID` and `ReferenceID`.
 - `species` -- Species. Links to `values` through `AccSpeciesID`.
}
